### 2023-06-02 #5515 @editorialbot
[PRE REVIEW]: GIRFReco.jl: An open-source pipeline for spiral MRI reconstruction in Julia
**Submitting author:** @alexjaffray (<a href="http://orcid.org/0000-0002-9571-1838">Alexander Jaffray</a>)
**Repository:** https://github.com/BRAIN-TO/GIRFReco.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
