### 2023-06-26 #5515 @jbytecode
I am a JOSS editor, and I am here to help you make your Julia package canonical before taking any further action. When I followed the README.md file in the software repository, I noticed a set of installation instructions that are not considered canonical for a Julia package. It doesn't even include a test folder and unit tests.

Could you please:

- Add a test folder and implement tests. You can refer to the documentation for testing capabilities if you are unsure: https://docs.julialang.org/en/v1/stdlib/Test/#Unit-Testing.
- Register the Julia package on JuliaHub so that users can install the package using the Pkg package in Julia: [docs](https://github.com/JuliaRegistries/Registrator.jl).
- Add GitHub Actions for online continuous integration: Here is an [example](https://github.com/jbytecode/JMcDM/blob/main/.github/workflows/main.yml).
- Update the README file as follows:

To install the package, use the following commands:

```Julia
julia> using Pkg
julia> Pkg.add(pkgname)
```

or

```
julia> ]
(@v1.9) pkg> add pkgname
```

etc.

Please don't hesitate to ask me anything if you encounter any difficulties while applying any of the steps mentioned above.
