### 2023-07-05 2be3adb
add forced GC
