### 2023-07-06 009849a
Added compat entries for Julia and dependencies.
