+++ Project.toml
@@ -52,3 +52,48 @@ Waveforms = "cb13b1c6-351e-5134-b3ad-d6a530956a82"
 [extras]
 CPUSummary = "2a0fbf3d-bb9c-48f3-b0a9-814d99fd7ab9"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+
+[compat]
+julia = "1.9.1"
+AxisArrays = "0.4.6"
+DSP = "0.7.8"
+DelimitedFiles = "1.9.1"
+Dierckx = "0.5.3"
+Documenter = "0.27.24"
+FFTW = "1.7.1"
+FileIO = "1.16.1"
+Flux = "0.13.17"
+FourierTools = "0.4.2"
+HDF5 = "0.16.15"
+HTTP = "1.9.8"
+ImageBinarization = "0.2.9"
+ImageEdgeDetection = "0.1.7"
+ImageMorphology = "0.3.2"
+ImageUtils = "0.2.9"
+Images = "0.25.3"
+LinearOperators = "2.5.2"
+Literate = "2.14.0"
+MAT = "0.10.5"
+MRIBase = "0.3.3"
+MRICoilSensitivities = "0.1.1"
+MRIFieldmaps = "0.0.3"
+MRIFiles = "0.1.6"
+MRIGradients = "0.1.0"
+MRIOperators = "0.1.2"
+MRIReco = "0.7.1"
+Measures = "0.3.2"
+MosaicViews = "0.3.4"
+NFFT = "0.13.3"
+NFFTTools = "0.2.6"
+NIfTI = "0.5.9"
+PlotlyJS = "0.18.10"
+Plots = "1.38.16"
+PolygonInbounds = "0.2.0"
+ROMEO = "1.0.1"
+RegularizedLeastSquares = "0.9.1"
+Scratch = "1.2.0"
+SparsityOperators = "0.4.7"
+StatsBase = "0.33.21"
+Unitful = "1.14.0"
+Waveforms = "0.1.1"
+Statistics = "1.9.0"
