+++ Project.toml
@@ -54,7 +54,7 @@ CPUSummary = "2a0fbf3d-bb9c-48f3-b0a9-814d99fd7ab9"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 
 [compat]
-julia = "1.9.1"
+julia = "1.9"
 AxisArrays = "0.4.6"
 DSP = "0.7.8"
 DelimitedFiles = "1.9.1"
@@ -78,7 +78,7 @@ MRIBase = "0.3.3"
 MRICoilSensitivities = "0.1.1"
 MRIFieldmaps = "0.0.3"
 MRIFiles = "0.1.6"
-MRIGradients = "0.1.0"
+MRIGradients = "0.1.1"
 MRIOperators = "0.1.2"
 MRIReco = "0.7.1"
 Measures = "0.3.2"
