+++ Project.toml
@@ -78,7 +78,7 @@ MRIBase = "0.3.3"
 MRICoilSensitivities = "0.1.1"
 MRIFieldmaps = "0.0.3"
 MRIFiles = "0.1.6"
-MRIGradients = "0.1.1"
+MRIGradients = "0.1.2"
 MRIOperators = "0.1.2"
 MRIReco = "0.7.1"
 Measures = "0.3.2"
