### 2023-07-07 #5515 @Kevin-Mattheus-Moerman
@alexjaffray I have just removed the `query-scope` label as based on the functionality this package offers this may be in scope. I did just now assign the `paused` label, just to flag that we will not resume until you've addressed the above. Please keep us up to date on progress, and if you can provide an estimate as to how long you think it will take, please do let us know.
