### 2023-07-21 #5515 @alexjaffray
@Kevin-Mattheus-Moerman Hi, we have just package-ized the project.

The package and paper detail a pipeline, for which the integration test is the successful reconstruction of non-Cartesian MR imaging data. We provide details about how to carry out the integration test in the README and the Literate example, but please let us know if you need anything else regarding testing.
