### 2023-07-29 #5515 @jbytecode
I can now install the package using Julia package manager by typing

```Julia
] add GIRFReco
```

however, when I am trying to perform the unit tests by typing

```Julia
] test GIRFReco
```

I get an error, because the package doesn't have tests. Referees will already remind about the absence of unit tests during the review process and maybe we can pass this for now, right? @Kevin-Mattheus-Moerman

Additionally, you can update the README.md file of repository with new installation instructions by replacing manual installation by automatic package installation.

A note on the GitHub integration tests: Since your software hasn't tests, CI does not perform a full check of your software, and it is still absent.

By the way, we have a registered package, and I believe all of the issues will be addressed by the reviewers, and the author will eventually correct them. We can also remind them to focus on software testing.
