### 2023-08-01 #5515 @alexjaffray
@jbytecode Thanks for the feedback, I will incorporate the suggestions and work on the testing of the package, I think the reviewers' input on a testing strategy will be very helpful.
