### 2023-08-01 58b4184
Update README.md
Update readme to include automatic package install instructions
