### 2023-08-03 cb50cbf
Update Julia version in README.md
