### 2023-08-03 #5515 @alexjaffray
@jbytecode I've added the canonical installation instructions for the package, and we want to hear reviewer feedback on appropriate tests for this package. Feel free to proceed.
