### 2023-08-03 #5515 @jbytecode
Looks great now. However, the package still lacks unit tests. I believe this issue will be emphasized in the review.

@Kevin-Mattheus-Moerman - I think this submission seems quite ready for review.
