### 2023-09-11 #5515 @Kevin-Mattheus-Moerman
@JeffFessler, @JakobAsslaender, @grlee77, @mathieuboudreau, @spinicist, @Laura2305, @timholy  would you be interested in reviewing the following submission for the [Journal of Open Source Software (JOSS)](https://joss.theoj.org). The submission is entitled: "GIRFReco.jl: An open-source pipeline for spiral MRI reconstruction in Julia"?

As you may know JOSS reviews focus on the the evaluation of the [the software](https://github.com/BRAIN-TO/GIRFReco.jl), and the review of a [short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05515/joss.05515/10.21105.joss.05515.pdf).

Please let me know if you are interested by responding here. Thanks!

JOSS reviews take place on GitHub using a streamlined process, for more information see our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). You can also look at this completed review process to see what the process looks like: https://github.com/openjournals/joss-reviews/issues/1725.
