### 2023-09-14 #5515 @cncastillo
Hi, thanks for considering for this. I would be glad to review this article.

Cheers!

Get Outlook for Android<https://aka.ms/AAb9ysg>
