### 2023-09-15 #5515 @mathieuboudreau
> I've never coded in Julia and though I'm very experience in MRI (14+ years), I've not dealt tooo much with image recon aglorithms (and not spiral, besides basic theory). But, if that's ok, I'm happy to review

Guess that's a no from you to me?
