### 2023-09-15 #5515 @mrikasper
> > I've never coded in Julia and though I'm very experience in MRI (14+ years), I've not dealt tooo much with image recon aglorithms (and not spiral, besides basic theory). But, if that's ok, I'm happy to review
>
> Guess that's a no from you to me?

@mathieuboudreau, thank you very much for offering to review, much appreciated. Maybe your modesty/disclaimer was dissuading @Kevin-Mattheus-Moerman ? Personally, I would be interested to see how a Julia-newbie would be able to use the tool, and you could dig deep into lack of documentation etc., but I am not sure what JOSS' expectations are exactly.
