### 2023-09-20 #5515 @korbinian90
I found a conflict of interest, as Lars Kasper and I are both co-authors on the [Neurodesk](https://pubmed.ncbi.nlm.nih.gov/36993557/) paper (50 authors in total). We didn't interact directly when contributing to Neurodesk however.
