### 2023-09-21 #5515 @felixhorger
I am interested! I have worked with GIRFs a couple of years ago. Glad to see more and more julia packages pop up in the MRI community :)
