### 2023-09-27 2dfd516
Update dev env descriptions.
