### 2023-09-29 BRAIN-TO/GIRFReco.jl#7 @felixhorger
Vector graphics
Any chance we can get [the in vivo image in the readme](https://github.com/BRAIN-TO/GIRFReco.jl/blob/9f5b4b98625a5aea1e87d77ca5505ed42ea13c10/README.md?plain=1#L59) in higher resolution or even better as a vector graphic? 😄 I think it would then be easier to appreciate how well your reconstruction works.
Take a look [here](https://stackoverflow.com/questions/13808020/include-an-svg-hosted-on-github-in-markdown).
