### 2023-09-29 #5877 @felixhorger
The DOIs it found for the ISMRM abstracts actually work, amazing, I didn't know about this 👍
@alexjaffray I think it's worth adding them into the .bib
