### 2023-09-29 #5877 @felixhorger
Please find my comments on the manuscript here: [10.21105.joss.05877_FH.pdf](https://github.com/openjournals/joss-reviews/files/12767854/10.21105.joss.05877_FH.pdf).

The main point I'd like to bring up is that the paper makes it seem like there isn't a clear separation in the code of the (abstract) pipeline itself, reading and writing of data and additional features like plotting. In my opinion there should be a clear separation, outsourcing as much as possible (I'm a fan of an ecosystem of packages rather than a toolbox). This separation would make use more flexible, and adaptation to new scenarios easier (you also mention this point in your paper). How exactly this could look like I'll have to clarify after I read the code in more detail.
