### 2023-09-30 BRAIN-TO/GIRFReco.jl#9 @felixhorger
Config files and root directory 
[This](https://github.com/BRAIN-TO/GIRFReco.jl/blob/9f5b4b98625a5aea1e87d77ca5505ed42ea13c10/docs/lit/examples/ReconConfig_joss_demo.jl#L151-L151) fails if your root directory isn't the main repo's root (that's one reason I brought up the IDE in #3, the IDE "hides" the fact that you're still running code from the repo's root. If you for example work with cmdline, a reasonable thing to do is to `cd` to the examples folder, and run the script from there, which doesn't work as described before.

I think what you should do is to make the example self-contained in a sense that it imports all packages it needs (`import` and `using`), and the rest of the code is available in the examples folder. This also ties in with my comments on the manuscript [here](https://github.com/openjournals/joss-reviews/issues/5877#issuecomment-1741110216): wouldn't it be better to have a function providing you with a dictionary (or other data structure) containing the options, and then in your example script you can adjust them to your need, instead of having to copy a file?
Open for discussion!
