### 2023-10-03 BRAIN-TO/GIRFReco.jl#7 @alexjaffray
@mrikasper can you re-generate a high resolution or svg version of the in vivo figure?
