### 2023-10-05 BRAIN-TO/GIRFReco.jl#7 @mrikasper
Thank you for the suggestion, @felixhorger !

The [source image](https://github.com/BRAIN-TO/GIRFReco.jl/blob/9f5b4b98625a5aea1e87d77ca5505ed42ea13c10/paper/figs/Fig_3.png) is already high resolution (in fact, higher resolution than the reconstruction matrix size), so I am not sure how JOSS for the paper PDF handles the conversion. If you know of any pointers how to make the conversion less lossy, I would be happy to implement it.

For the README, we use a [lower-resolution version](https://github.com/BRAIN-TO/GIRFReco.jl/blob/main/docs/figs/In-vivo-demo.png) of the image to reduce loading time (670kB instead of 9MB), but its resolution is still close to the actual resolution of the reconstruction matrix size.

With respect to vector graphics formats, if I understand them correctly, they would not improve the quality of the pixel-based reconstruction images themselves, but only the text and lines in between. File size might be a bit smaller, but in my opinion accessibility is higher with png, because many programs and platforms can read them.

All the best,
Lars
