### 2023-10-06 BRAIN-TO/GIRFReco.jl#7 @felixhorger
Hi Lars! That's a good point with the size for the readme and accessibility.
Like you say, using vector graphics would improve text and annotation, but not do anything for the images. I brought this up because I think it's good practice in science and if set up correctly can prevent the resolution problem to happen. But this is a personal preference, so if you want to stick to `png` that's fine by me ;)

Let me elaborate, because my previous comment was indeed not precise enough, sorry for that.
I tried to say that the nominal resolution is not the effective resolution in the images, giving them a blurry appearance if you zoom in (even with a viewer that doesn't interpolate). This is because the images were interpolated beforehand I assume.
Take a look at this screenshot made from ms-paint (which doesn't interpolate):

There you can see the pixels, but the effective resolution is worse than that (the nominal image resolution is clearly sub-millimeter, but the nominal resolution of the scan is probably 1mm?). You can specifically see it for the red line, which once was 1 pixel wide I suppose, before interpolation.

I think the resolution of the presented data should equal the nominal resolution of the scan, i.e. data is presented without any directly necessary post-processing.

Regarding how I would go about making the figure, something like:
```
fig, axs = PyPlot.subplots(2, 5)
axs[1,1].imshow(your_image; interpolation="none")
...
# use axs[2, 4] for a colourbar for the difference maps? and axs[2, 5] for the Ref Edge plot
PyPlot.savefig("myfig.png"; dpi=200) # dpi value not tested, can be lower to save memory, or higher to prevent misalignment
```
Adding the braces is possible I guess, but difficult. I suggest instead using the titles `No correction (A)`, `+ B0 map (B)`, `+GIRF Kxyz (C)`, ... and the titles of the difference maps `|A - B| x 5` etc.

Thanks and best wishes,
Felix
