### 2023-10-17 #5877 @alexjaffray
> Please find my comments on the manuscript here: [10.21105.joss.05877_FH.pdf](https://github.com/openjournals/joss-reviews/files/12767854/10.21105.joss.05877_FH.pdf).
>
> The main point I'd like to bring up is that the paper makes it seem like there isn't a clear separation in the code of the (abstract) pipeline itself, reading and writing of data and additional features like plotting. In my opinion there should be a clear separation, outsourcing as much as possible (I'm a fan of an ecosystem of packages rather than a toolbox). This separation would make use more flexible, and adaptation to new scenarios easier (you also mention this point in your paper). How exactly this could look like I'll have to clarify after I read the code in more detail.

Thanks for the feedback! Your pdf comments seem to evolve over the course of the document, and you mentioned that you will read the code and clarify some things. Should we start to address your comments, or should we wait until you are finished (preferred)? We are also very happy to address each code review issue dynamically in Github, but likely it would be good to have all reviewers comments and suggestions first so that we don't duplicate work. Would this be alright with you @felixhorger
