### 2023-10-18 #5877 @felixhorger
@alexjaffray yes I think that's a good idea, there are some minor comments on clarity which could be implemented already so that the other reviewers can proof-read that. The comments on the structure of the code can be left aside for now until all reviewers finished reading the code and gave their opinion.
