### 2023-10-18 #5877 @Kevin-Mattheus-Moerman
@alexjaffray @felixhorger you may choose the approach that you like. However, usually we see the review progresses fastest/smoothest when the authors respond to issues directly as soon as they come in. This way the reviewers do not have to wait (their issues are being dealt with quickly).
