### 2023-10-19 BRAIN-TO/GIRFReco.jl#11 @mathieuboudreau
Add more installation details for users new to Julia
Specifically, this instruction,

https://github.com/BRAIN-TO/GIRFReco.jl/blob/9f5b4b98625a5aea1e87d77ca5505ed42ea13c10/README.md#L15-L19

was not clear to me as someone with only a little Julia experience (and last time I used it was a few years ago).

Because, "use the following command" usually means "copy and paste this" into the terminal/software, and this command as is doesn't work in neither the mac Terminal or the Julia session. Even copy pasting `]add GIRFReco` doesn't work once inside the Julia terminal; you need to explicitely **type** it because typed `]` opens the julia package manager but copy-pasting "]" with a command doesn't.

I'd recommend breaking it down into three instructions.

1. Open a Julia session
2. Open the package manager by typing `]`
3. Add the package using the command `add GIRFReco` once in the package manager.
