### 2023-10-19 #5877 @mathieuboudreau
I started going through the checklist. Here I'll list issues in the repo that I've opened and if they've been completed/closed or not yet.

- [ ] Installation clarity - https://github.com/BRAIN-TO/GIRFReco.jl/issues/11
