### 2023-10-19 BRAIN-TO/GIRFReco.jl#12 @cncastillo
Reduction of strong dependencies in Project.toml
I think the number of strong dependencies in the package needs to be reduced to a minimum to avoid compatibility problems, **based on what you want the core package to do**. As @felixhorger mentioned, one way of doing this is to separate the package into various simpler packages with minimal dependencies (which I am a fan of!). Another way of doing this, which is easier to implement, is to use [package extensions](https://pkgdocs.julialang.org/v1.9/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions)).

Having said that, there are some packages that ultimately **should not** be in the Project.toml of the main package, like `Literate.jl` and `Documenter.jl` as these correspond to the Project.toml of the docs (the same for the tests, and the generation of the figures for the paper). This comment also applies to `MRIGradients.jl` (whose name is a little confusing because it only appears to do GIRF correction).

The main point that I want to come across is that if, as a user, I want to do only a GIRF-corrected recon, having my k-space, GIRF, and coils estimated my way, I do not expect a package called `GIRFReco.jl` to use DL tools like `Flux.jl` for B0 estimation and coil estimation, or to require `Nifti.jl`, `MAT.jl`, `Unitful.jl`, and others. In the specific case of Flux.jl, it may not be needed at all, and the gradient could be computed manually (I only saw it being used in `pcg_ml_est_fieldmap` and `sens_smooth`). Arguably,  functions that do B0 and coil sensitivity estimation should not be part of this package at all.

Going back to the package extensions, the idea is to define a function only if the package is explicitly imported and added as a weak dependency. A simple example is the function `readGIRFFile` (from `MRIGradients.jl`) that depends on `MAT.jl` (for some reason, `MAT.jl` is also a dependency of `GIRFReco.jl`?).  Only if the user does this
```julia
using MRIGradients, MAT
```
Then you get the function `readGIRFFile` defined in `/ext/MRIGradientsMATExt.jl` as
```julia
module MRIGradientsMATExt
    import MRIGradients
    using MAT
    # definition of readGIRFFile with more arguments, inside MRIGradients the function
    # should be defined as an empty function: function readGIRFFile end
    MRIGradients.readGIRFFile( ...)
        ...
    end
end
```
that should also be added to the Project.toml
```toml
[weakdeps]
MAT = "..."

[extensions]
MRIGradientsMATExt = "MAT"
```

For the case of Flux.jl, if it is truly necessary, I think the best way of doing this would be to add Flux as a weak dependence of `MRIFieldmaps.jl` and `MRICoilSensitivities.jl`. In that way, if I want to do a field map estimation, I do
```julia
using MRIFieldmaps
# My code that uses MRIFieldmaps
data = ... # read data somehow
B0 = estimate_field_map(data, ..) #It is probably not called like this
```
and if I want to use the functions defined with Flux
```julia
using MRIFieldmaps, Flux
# My code that uses MRIFieldmaps
data = ... # read data somehow
B0 = estimate_field_map(data, ..) #I can estimate B0 the "traditional way"
B0 = estimate_field_map_flux(data, ..) #But also using Flux for the gradient!
```

Hope this is useful,
Cheers!
