### 2023-10-19 BRAIN-TO/GIRFReco.jl#12 @mrikasper
Dear Carlos @cncastillo,

Thank you so much for this detailed feedback, we will go through the dependencies carefully and see which ones can be dropped/transformed into weak dependencies.
