### 2023-10-19 #5877 @mrikasper
Dear Felix @felixhorger and Carlos @cncastillo,

I noticed both in [Felix' review comment here](https://github.com/openjournals/joss-reviews/issues/5877#issuecomment-1741110216) and [one of Carlos' issues](https://github.com/BRAIN-TO/GIRFReco.jl/issues/12) a more conceptual question that I wanted to discuss a bit: Both of you are emphasizing the importance of modularization and small packages with limited, but efficient functionality.

I understand the need for that and the elegance in replacing certain tools based on one's specific needs, but in the end, in order to solve a complex task, we have to combine these tools into an elaborate pipeline that needs to make specific choices and adapt to the idiosyncrasies of our use cases. Reconstructing simulated data from a nominal Archimedean spiral is something different than turning real data from a real scanner into images that should then be used in established neuroimaging tools for further analysis, like FSL DTIFIT for diffusion (which rely on NIfTI as input format).

This is what our submission is mainly about (_a pipeline for..._), in particular `GIRFReco.jl`, whereas `MRIGradients.jl` is more of a tool in the sense that you describe it. In other parts of neuroimaging, having complex pipelines standardized and published is definitely seen as valuable, for example [fmriprep](https://www.fmriprep.org) for functional MRI, and I don't see why image reconstruction would differ in this respect.

Is this maybe a Julia-specific preference? Or because of what the notion of a "package" is in Julia? Originally, our repository was not even a package, but rather a project, and I would say mainly for practical reasons, e.g., ease of installation, we changed this during the pre-review process. It does make sense to have it as a package, I think, because other deployment features also become standardized (e.g., documentation and testing, which you pointed out), but maybe this is where the different expectations come from?

Looking forward to your insights,
Lars
