### 2023-10-19 BRAIN-TO/GIRFReco.jl#7 @mrikasper
Dear Felix @felixhorger,

Thank you for this detailed response and the example code.

I understand now better what you suggest: Let the reader decide on what interpolation is used for displaying a pixel matrix, rather than the author.

I think ideally this is done on a programmable level like you suggest with your coding snippets. However, our ethics does not allow to share the data for the in-vivo figure at the moment, so this would not really be possible in and end-to-end way.

What we could do instead is to do what you suggest for the phantom figure and save it (or its components) as `.svg`. Does this sound like a good compromise?

All the best,
Lars
