### 2023-10-20 #5877 @cncastillo
Dear Lars @mrikasper,

To keep myself from overextending here, I will reply about modularization in BRAIN-TO/GIRFReco.jl#12.

I think there could be a naming issue, as the name of the package `GIRFReco.jl` gives the impression that it provides the tools to do a GIRF-corrected reconstruction from the user's data (which is conceptually different from the coils sens. and B0 map estimation). For example, if I want to include the linear gradient self-terms and $k_0$ corrections to my MRF radial data (or spiral), it seems that the most useful package would be `MRIGradients.jl`, as I could not easily use `GIRFReco.jl`.

A name that more closely represents what the packages are meant to do (core functionality) is needed to set expectations correctly. For example, `MRISpiralDiffusionReco.jl`, and `MRIGIRFCompensation.jl`. Maybe the efforts in documentation, tests, CI, etc, need to go to the package currently named `MRIGradients.jl`, and the presented pipeline can be an example use case that integrates fancier coils estimation, B0 correction, etc.

Cheers!
