### 2023-10-21 BRAIN-TO/GIRFReco.jl#11 @felixhorger
Good point, even though I'd recommend not expanding the help here, because if every package does that for every of the installation steps it's longer to read for the ones who know how it works (+repetition). How about putting a link to the official julia manual, there is a chapter on the package manager. That would be a one-liner and point beginners in the right direction.

Here
https://github.com/BRAIN-TO/GIRFReco.jl/blob/9f5b4b98625a5aea1e87d77ca5505ed42ea13c10/README.md?plain=1#L16-L19

I could imagine something like `[To install the package from the julia REPL](https://link.to.manual/packagemanager) use:`
