### 2023-10-21 BRAIN-TO/GIRFReco.jl#12 @felixhorger
> First of all, I just want to clarify that this package has done a fantastic job, and I understand what a thankless job it is to publish an open-source package. Making open-source software useful for the community requires some painful, and not really publishable, laborious steps: documentation, testing, following the language guidelines, etc.

I can only repeat that! :)
