### 2023-10-21 BRAIN-TO/GIRFReco.jl#11 @mathieuboudreau
Thats fair! I think at a minimum, if hou’re going to say ] and the command in the same line, the instructions should be to type it (to avoid copy-pasting issue like I had, very few ppl in MRI are using Julia atm, see my editorial in MRM https://onlinelibrary.wiley.com/doi/10.1002/mrm.29366)
