### 2023-10-21 BRAIN-TO/GIRFReco.jl#7 @felixhorger
Dear Lars,

thanks for elaborating! I think the phantom scan is good in terms of the nominal resolution being equal to the effective one (when I zoom-in to the pixel level there is no blurring visible). I can for example spot the one-pixel-wide intensity modulations around the edges of the phantom internals in the second subplot of the phantom figure. If you compare that to the snapshot I put above you'll notice that the pixels are more fine grained than the structures in the image, i.e. the effective resolution is worse than the nominal.

My proposition does not need public access to the data or any interaction with the user/reader. To clarify what I think happened: the images were reconstructed, plotted into pngs (possibly first interpolation), then imported into another program (PowerPoint/Word?) to add the brackets and labels (possibly second interpolation) and then exported as a pdf or png (possibly third interpolation). What I would like is to make sure all these steps do nearest neighbour interpolation instead of bilinear (or similar). Let me provide another snippet which shows an exaggerated example, hopefully supporting my point better
```julia
using PyPlot
plt.ioff()

image = zeros(8, 8)
image[:, 4] .= 1
image[4, 5] = 2

fig, axs = plt.subplots(1, 2, figsize=(7, 3))
axs[1].imshow(image, interpolation="bilinear")
axs[2].imshow(image, interpolation="nearest")
foreach(ax -> ax.set_axis_off(), axs)
plt.savefig("test.png", dpi=20)
```
and then you open `test.png` in ms-paint, rescale by 400%, which should give this

showing that on the left, where linear interpolation is used, a blurring is introduced, and the effective resolution is obviously worse than the nominal one. On the right, the image matrix still has the same size as on the left, but it's obvious what the effective resolution is, i.e. I can make out the original pixels of the 8x8 data `image`.

Apologies for dragging this out, but I think the power of your pipeline would be more appreciable if you can see the same level of detail you have in the phantom scans also in the in vivo images. I see this problem with interpolation of lower-res images quite often in publications or ISMRM abstracts, so I want to develop a best practice for displaying images in our field.

All the best,
Felix
