### 2023-10-22 BRAIN-TO/GIRFReco.jl#11 @alexjaffray
@mathieuboudreau There is an explicit Julia-n way to do this which avoids the copy paste confusion. Some packages include it, but we elected not to. I will add it for clarity. The copy and paste friendly version is:

```
using Pkg
Pkg.add("GIRFReco")
```
