### 2023-10-22 #5877 @felixhorger
Regarding the reproducibility of original results from the paper: there are phantom and in vivo results. The in vivo results cannot be reproduced by the reviewers due to data protection (see @mrikasper's [comment here](https://github.com/BRAIN-TO/GIRFReco.jl/issues/7#issuecomment-1771770980)).
For the phantom results, there is no code available, maybe it makes sense to include that in the `joss_example.jl`, i.e. to produce exactly the same plot as in the paper on top of the existing ones?
