### 2023-10-22 #5877 @felixhorger
I reckon this point
> - [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

is implicitly fulfilled because the code is on github, hence anyone can do 1), 2), 3)? Or is an explicit statement required @Kevin-Mattheus-Moerman?
