### 2023-10-22 BRAIN-TO/GIRFReco.jl#11 @cncastillo
Hi, maybe I can give my grain of salt on this. If you copy and paste this line in the Julia REPL
```julia
(@v1.9) pkg> add GIRFReco
```
It just works :sunglasses: . At least in Linux and Mac, not quite sure about Windows. I agree with @mathieuboudreau
 that the word **type** is fundamental for the `] add GIRFReco` instruction. In our experience with students with 0 Julia experience saying "type `]add Package` in the Julia REPL" was enough for them to install the package successfully.

As @felixhorger said, I would not put so much information in the `README.md`, and would refrain from using `Pkg` as it gives the impression to new users that it is more complicated than it actually is.

If you want to give additional instrucctions (nothing wrong with that), you could add a [Getting Started](https://cncastillo.github.io/KomaMRI.jl/stable/getting-started/) section in the documentation
