### 2023-10-22 BRAIN-TO/GIRFReco.jl#9 @felixhorger
Same problem here

https://github.com/BRAIN-TO/GIRFReco.jl/blob/9f5b4b98625a5aea1e87d77ca5505ed42ea13c10/docs/lit/examples/joss_demo.jl#L102-L102

are there reasons for not making it a function? You already mentioned that historically `GIRFReco.jl` wasn't a package during development, but now that it is I see no reason why this functionality should remain in a script.
