### 2023-10-22 BRAIN-TO/GIRFReco.jl#12 @felixhorger
Relevant to this is also @mrikasper's [comment](https://github.com/openjournals/joss-reviews/issues/5877#issuecomment-1771758595) in the JOSS review

> This is what our submission is mainly about (*a pipeline for...*)

I think a reasonable thing to do is to modularise your pipeline into abstract steps, potentially outsource into separate packages (e.g. the B0 field estimation), and provide an overarching function which starts the whole pipeline, given the dictionary of parameters. It can then read from ISMRMRD and write into NIFTI files, exactly as it is now. I recommend to add one version of that function where arrays are provided instead of reading them from file, and arrays are returned instead of writing to NIFTI. Further, I recommend that all subparts of the pipeline follow the same principle of working on arrays rather than files.

With this setup, beginners are happy because they have a plug and play version and advanced users are happy because it's modular and they can exchange parts of the pipeline and are not forced to provide data in a specific format or always write to disk. Happy to hear your opinions on this :)
