### 2023-10-23 #5877 @Kevin-Mattheus-Moerman
@alexjaffray @felixhorger in terms of the "Community guidelines". We recommend that the authors create a dedicated `CONTRIBUTING.md` file in which the guidelines to contribute are spelled out. This file should be found in the main folder are should be linked to from the `README.md`. This website has some nice examples:  https://contributing.md/example/.

@alexjaffray if you find the dedicated file "overkill" you can also instead have a simple section in your README where you outline how you wish folks to contribute to the project etc. So similar content but shorter. We leave this up to you.
