### 2023-10-24 daa1608
update readme with clear type vs copy instruction
