### 2023-10-24 06c5e81
fix quick-installation link in README
