### 2023-10-24 aff10d5
re-order instructions to be more natural
