### 2023-10-24 e81f4e4
fix fig number and remove custom path
