### 2023-10-24 #5877 @alexjaffray
@Kevin-Mattheus-Moerman Thanks for the suggestion of adding community contributing guidelines, I'll take a look at some examples and will add a dedicated file, I think this will be helpful and make the README less crowded
