### 2023-10-24 38e0ec5
add automatic TE reading
