+++ recon/CartesianRecon.jl
@@ -12,7 +12,7 @@ include("../utils/fieldMapEstimator.jl")
 b0FileName = paramsGeneral[:fullPathMapScan];
 
 # filename for preprocessed data (remove oversampling, permute dimensions wrt MRIReco)
-processedFileName = paramsGeneral[:fullPathProcessedMapScan] 
+processedFileName = paramsGeneral[:fullPathProcessedMapScan]
 
 if paramsGeneral[:doProcessMapScan]
 
@@ -35,10 +35,10 @@ dataFileNew = ISMRMRDFile(processedFileName)
 
 
 rawDataNew = RawAcquisitionData(dataFileNew)
-acqDataCartesian= AcquisitionData(rawDataNew, estimateProfileCenter=true)
+acqDataCartesian = AcquisitionData(rawDataNew, estimateProfileCenter = true)
 
 # Define coils and slices
-nCoils = size(acqDataCartesian.kdata[1],2)
+nCoils = size(acqDataCartesian.kdata[1], 2)
 nSlices = numSlices(acqDataCartesian)
 
 # Define TEs 
@@ -59,7 +59,7 @@ shiftksp!(acqDataCartesian, paramsGeneral[:fovShift]) # amount of FOV shift; in 
 @info "Calculating Sense Maps"
 
 # from reading docstring, min thresholding
-senseCartesian = espirit(acqDataCartesian,(6,6),30,eigThresh_1=0.01, eigThresh_2=0.9)
+senseCartesian = espirit(acqDataCartesian, (6, 6), 30, eigThresh_1 = 0.01, eigThresh_2 = 0.9)
 
 # from Lars (7T spirals)
 #senseCartesian = espirit(acqDataCartesian,(6,6),30,eigThresh_1=0.02, eigThresh_2=0.98)
@@ -68,17 +68,17 @@ senseCartesian = espirit(acqDataCartesian,(6,6),30,eigThresh_1=0.01, eigThresh_2
 
 # normalize for consistency with saving/loading and better ranges of reconstruction values
 senseCartesian /= maximum(abs.(senseCartesian))
-sensitivity = senseCartesian 
+sensitivity = senseCartesian
 
-res_x = fieldOfView(acqDataCartesian)[1]./ size(sensitivity)[1]
-res_y = fieldOfView(acqDataCartesian)[2]./ size(sensitivity)[2]
-res_z = fieldOfView(acqDataCartesian)[3] .*(1 + paramsGeneral[:sliceDistanceFactor_percent]./100.0) # for 2D only, since FOV[3] is slice thickness then, but gap has to be observed
+res_x = fieldOfView(acqDataCartesian)[1] ./ size(sensitivity)[1]
+res_y = fieldOfView(acqDataCartesian)[2] ./ size(sensitivity)[2]
+res_z = fieldOfView(acqDataCartesian)[3] .* (1 + paramsGeneral[:sliceDistanceFactor_percent] ./ 100.0) # for 2D only, since FOV[3] is slice thickness then, but gap has to be observed
 resolution_mm = (res_x, res_y, res_z)
 
 # save SENSE maps
 if paramsGeneral[:doSaveRecon] # TODO: include elements to save as tuple, e.g., ["b0", "sense", "recon"], same for load
     # TODO: use correct slice order everywhere, e.g., when saving/loading maps for spiral recon
-    saveMap(paramsGeneral[:fullPathSaveSense], sensitivity[:,:,sliceIndexArray,:], resolution_mm; doSplitPhase=true)
+    saveMap(paramsGeneral[:fullPathSaveSense], sensitivity[:, :, sliceIndexArray, :], resolution_mm; doSplitPhase = true)
 end
 
 ## Parameter dictionary definition for reconstruction
@@ -86,52 +86,61 @@ end
 @info "Setting Parameters"
[…]

[…]
