### 2023-10-30 b18ebda
cleanup moved legacy scripts
