+++ recon/SpiralRecon_Cleaned_wavelet.jl
@@ -1,151 +1,0 @@
-using PyPlot, HDF5, MRIReco, LinearAlgebra, Dierckx, DSP, FourierTools, ImageBinarization, ImageEdgeDetection, MRIGradients
-
-# %%
-# Include tools and reader functions for running the spiral reconstruction recipe
-# Note: the files are found relative of the location of the folder, not the
-# environment current folder
-include("../io/GradientReader.jl")
-include("../utils/Utils.jl")
-
-## Executing Cartesian recon from which B0/sensitivity maps have been computed
-@info "Running julia_recon_cartesian to retrieve maps (senseCartesian and b0Maps)"
-include("../recon/CartesianRecon.jl")
-
-## Set figures to be unlocked from the window (i.e use matplotlib backend with controls)
-pygui(true)
-
-## Choose Slice (can be [single number] OR [1,2,3,4,5,6,7,8,9]
-# sliceChoice = [1,2,3,4,5,6,7,8,9] # UNCOMMENT FOR MULTISLICE
-sliceChoice = [3] # UNCOMMENT FOR SINGLESLICE (SLICES 3, 7 and 8 are good examples)
-diffusionDirection = 0 # CAN BE FROM 0 (b=0) to 7 (1-7 are 6 directions of b=1000)
-
-## Spiral Reconstruction Recipe Starts Here
-@info "Starting Spiral Reconstruction Pipeline"
-
-## Default to single slice selection. Choose multi-slice only if computer is capable.
-multiSlice = false
-
-if length(sliceChoice) > 1
-    multiSlice = true
-end
-
-if !multiSlice
-    selectedSlice = sliceChoice
-else
-    selectedSlice = sort(vec(sliceChoice))
-end
-
-## The ISMRMRD File contains more than one excitation, so we choose the set corresponding to the b-value 0 images
-excitationList = vec(20:2:36) .+ diffusionDirection * 18 # DATASET SPECIFIC INDEXING
-sliceSelection = excitationList[selectedSlice]
-
-@info "Slice Chosen = $selectedSlice: \n \nExcitations Chosen = $excitationList "
-
-# adjustmentDict is the dictionary that sets the information for correct data loading and trajectory and data synchronization
-adjustmentDict = Dict{Symbol,Any}()
-adjustmentDict[:reconSize] = (200, 200)
-adjustmentDict[:interleave] = 1
-adjustmentDict[:slices] = 1
-adjustmentDict[:coils] = 20
-adjustmentDict[:numSamples] = 15475
-adjustmentDict[:delay] = 0.00000 # naive delay correction
-
-adjustmentDict[:interleaveDataFileNames] = ["data/Spirals/523_96_2.h5", "data/Spirals/523_98_2.h5", "data/Spirals/523_100_2.h5", "data/Spirals/523_102_2.h5"]
-adjustmentDict[:trajFilename] = "data/Gradients/gradients523.txt"
-adjustmentDict[:excitations] = sliceSelection
-
-adjustmentDict[:doMultiInterleave] = false
-adjustmentDict[:doOddInterleave] = false
-adjustmentDict[:numInterleaves] = 1
-
-adjustmentDict[:singleSlice] = !multiSlice
-
-@info "Using Parameters:\n\nreconSize = $(adjustmentDict[:reconSize]) \n interleave = $(adjustmentDict[:interleave]) \n slices = $(adjustmentDict[:slices]) \n coils = $(adjustmentDict[:coils]) \n numSamples = $(adjustmentDict[:numSamples])\n\n"
[…]
