+++ recon/ReconConfig_SPIDI_0011.jl
@@ -1,140 +1,0 @@
-## This ReconConfig.jl file describes all reconstruction parameters, as well as data locations and selections for an iterative non-Cartesian reconstruction that relies 
-#  on an external reference scan (Cartesian) to estimate calibration maps (coil sensitivities, B0 maps)
-
-using Dates
-
-paramsGeneral = Dict{Symbol,Any}()
-
-## General options for recon script
-paramsGeneral[:doLoadMaps] = true               # if true, reloads B0/SENSE maps instead of recalculating
-paramsGeneral[:doSaveRecon] = true              # if true, saves reconstruction and all auxiliary image data (maps) as NIfTI files
-paramsGeneral[:doPlotRecon] = false             # if true, plots intermediate debugging and output recon figures (needs graphics, not recommended in multi-thread mode due to PyPlot)
-paramsGeneral[:doProcessMapScan] = true         # if true, compute sensitivity and B0 maps from reconstructed Cartesian scan   
-paramsGeneral[:doSaveProcessedMapScan] = false; # save ISMRMD file of preprocessed Cartesian data (before recon)
-
-## Reconstruction Parameters
-# update time stamp for new recon, otherwise keep fixed, will create a new recon/<reconId> directory
-#paramsGeneral[:reconId] = Dates.format(Dates.now(), "yyyy-mm-dd_HH_MM_SS") # recon ID is reconId
-# paramsGeneral[:reconId] = "2022-10-20_09_07_07"
-paramsGeneral[:reconId] = "v999";
-paramsGeneral[:doCorrectWithB0map] = true
-paramsGeneral[:doCorrectWithGIRFkxyz] = true
-paramsGeneral[:doCorrectWithGIRFk0] = true
-
-paramsGeneral[:nVirtualCoils] = 8;
-paramsGeneral[:doCoilCompression] = false;
-paramsGeneral[:fovShift] = [0, -10]; # Unit: number of voxels
-
-## Scan parameters, Additional acquisition information, e.g., slice distance etc.
-paramsGeneral[:sliceDistanceFactor_percent] = 000 # 400
-
-#Total number of ADC points BEFORE the rewinder at the end of the spiral readout. For gradient 508, use 15655 (out of 16084); for gradient 511, use 15445 (out of 15624).
-paramsGeneral[:numADCSamples] = 15655 # 15504
-# Matrix size of the reconstructed image. For gradient 508 with all 4 interleaves, use 200 for high resolution image; otherwise consider using 112 or 84 for a lower resolution. The FOV is 220 mm for both gradients 508 and 511.
-paramsGeneral[:reconSize] = (200, 200) #(112, 112) #(200, 200)
-paramsGeneral[:nReconIterations] = 20; # number of recon iterations (for both Cartesian and Spiral recon)
-paramsGeneral[:b0mapSmoothBeta] = 0.1 # for estimateB0Maps, * `β` - Regularization parameter controlling roughness penalty (larger = smoother, default 5e-4)
-paramsGeneral[:doNormalizeRecon] = false # set max abs to 1
-paramsGeneral[:scalingFactorSaveRecon] = 1.0e9 # 1 # typical range of recon intensities is 1e-7, rescale when saving, e.g., to 0...1000 roughly for fMRI analysis
-
-# Data selector
[…]
