### 2023-10-30 8fe9644
add test placeholder functions
