+++ test/runtests.jl
@@ -1,1 +1,4 @@
 using GIRFReco
+using Test
+
+include("testUtils.jl")
