+++ test/testIntegration.jl
@@ -1,0 +1,20 @@
+function test_integration_1()
+
+end
+
+function test_integration_2()
+
+end
+
+function test_integration_3()
+
+end
+
+function test_integration()
+    @testset "Integration" begin
+        test_integration_1()
+        test_integration_2()
+        test_integration_3()
+    end
+end
+
