### 2023-11-02 #5877 @alexjaffray
Thanks for the reminder, we’re working on all the other issues that have
been raised at the moment (and making good progress) but I’m happy to add a
contributing guide. I’ll add it to our internal project board and update
once it’s been pushed to our response branch.

On Thu, 2 Nov 2023 at 01:47, Kevin Mattheus Moerman <
***@***.***> wrote:

> @alexjaffray <https://github.com/alexjaffray> any updates on the above?
>
