### 2023-11-02 22bd728
WIP on renaming functions and vars
