+++ legacy/SpiralRecon_Cleaned_Mar2022_Human_SingleIntlv.jl
@@ -31,7 +31,7 @@ gamma = 42577478
 ## Choose diffusion direction; starting from 0 (b=0) to the total number in MDDW protocol, e.g. for 6 diffusion directions, 1-6 stands for 6 DWIs)
 diffusionDirection = selector[:dif]
 idxAverage = selector[:avg]
-nDiffusionDirections = paramsGeneral[:nDiffusionDirections] # TODO: Read from ISMRMRD itself
+nDiffusionDirections = params_general[:nDiffusionDirections] # TODO: Read from ISMRMRD itself
 
 # Which interleave to be reconstructed. For single-interleave data, it will always be set as 1; for multi-interleave data, the value set here will be used.
 # For multi-interleaved data, this value is ranging from [1:TotNumIntlv] (total number of interleaves), indicating which interleave to be reconstructed
@@ -39,25 +39,25 @@ startIndexIntlv = selector[:seg]
 
 
 ## Determine to reconstruct single-interleave data, or one interleave out of multi-interleave data.
-isDataSingleIntlv = isa(paramsGeneral[:fullPathScan], String)
+isDataSingleIntlv = isa(params_general[:fullPathScan], String)
 
 
 ## ------------------------------------------------ Calculation Starts Here ---------------------------------------------------------- ##
 
-if paramsGeneral[:doLoadMaps] && isfile(paramsGeneral[:fullPathSaveB0]) # # TODO ask for sense map (but split in magn/phase)
-    @info "Loading SENSE and B0 maps from $(paramsGeneral[:fullPathSaveSense]) and $(paramsGeneral[:fullPathSaveB0])"
+if params_general[:do_load_maps] && isfile(params_general[:fullPathSaveB0]) # # TODO ask for sense map (but split in magn/phase)
+    @info "Loading SENSE and B0 maps from $(params_general[:fullPathSaveSense]) and $(params_general[:fullPathSaveB0])"
     # load maps, permute slice, sice files have geometric slice order
-    b0Maps = loadMap(paramsGeneral[:fullPathSaveB0])
+    b0Maps = load_map(params_general[:fullPathSaveB0])
 
     nSlices = size(b0Maps, 3)
-    sliceIndexArray = getSliceOrder(nSlices, isSliceInterleaved = true)
+    sliceIndexArray = get_slice_order(nSlices, isSliceInterleaved = true)
 
     b0Maps = b0Maps[:, :, invperm(sliceIndexArray)]
-    senseCartesian = loadMap(paramsGeneral[:fullPathSaveSense]; doSplitPhase = true)[:, :, invperm(sliceIndexArray), :]
+    cartesian_sensitivity = load_map(params_general[:fullPathSaveSense]; doSplitPhase = true)[:, :, invperm(sliceIndexArray), :]
 else
     ## Only calculate sensitivity and B0 maps when they have not been done yet, or it's specifically required.
     ## Executing Cartesian recon from which B0/sensitivity maps have been computed
-    @info "Running CartesianRecon to retrieve maps (senseCartesian and b0Maps)"
+    @info "Running CartesianRecon to retrieve maps (cartesian_sensitivity and b0Maps)"
     include("CartesianRecon_Mar2022_Human.jl")
     nSlices = size(b0Maps, 3)
 end
@@ -88,15 +88,15 @@ sliceSelection = excitationList[selectedSlice]
 
[…]
