+++ legacy/SpiralRecon_Cleaned_Mar2022_Human.jl
@@ -16,9 +16,9 @@ doCoilCompression = true
 gamma = 42577478
 
 ## Only calculate sensitivity and B0 maps when they have not been done yet, or it's specifically required.
-if reloadCartesianData || !((@isdefined senseCartesian) && (@isdefined b0Maps))
+if reloadCartesianData || !((@isdefined cartesian_sensitivity) && (@isdefined b0Maps))
     ## Executing Cartesian recon from which B0/sensitivity maps have been computed
-    @info "Running CartesianRecon to retrieve maps (senseCartesian and b0Maps)"
+    @info "Running CartesianRecon to retrieve maps (cartesian_sensitivity and b0Maps)"
     include("CartesianRecon_Mar2022_Human.jl")
 end
 
@@ -90,7 +90,7 @@ if reloadSpiralData || !(@isdefined acqDataImaging)
     ## Convert raw to AcquisitionData
 
     @info "Merging interleaves and reading data \n"
-    acqDataImaging = mergeRawInterleaves(adjustmentDict)
+    acqDataImaging = merge_raw_interleaves(adjustmentDict)
 
     @info "Loading Gradient Impulse Response Functions \n"
     ## Load GIRFs!
@@ -100,7 +100,7 @@ if reloadSpiralData || !(@isdefined acqDataImaging)
     gAk1 = GirfApplier(gK1, gamma)
 
     @info "Correcting For GIRF \n"
-    applyGIRF!(acqDataImaging, gAk1)
+    apply_girf!(acqDataImaging, gAk1)
 
     # Load K₀ GIRF
     # Tim Wu, use new read GIRF function
@@ -109,10 +109,10 @@ if reloadSpiralData || !(@isdefined acqDataImaging)
     gAk0 = GirfApplier(gK0, gamma)
 
     @info "Correcting For k₀ \n"
-    applyK0!(acqDataImaging, gAk0)
+    apply_k0!(acqDataImaging, gAk0)
 
     ## Check the k-space nodes so they don't exceed frequency limits [-0.5, 0.5] (inclusive)
-    checkAcquisitionNodes!(acqDataImaging)
+    check_acquisition_nodes!(acqDataImaging)
 
 end
 
@@ -135,9 +135,9 @@ if doCoilCompression
 end
 
 # ## Plot the sensitivity maps of each coil
-if paramsGeneral[:doPlotRecon]
+if params_general[:do_plot_recon]
     @info "Plotting SENSE Maps"
-    plotSenseMaps(sensitivity, nvcoils, sliceIndex = 10)
+    plot_sense_maps(sensitivity, nvcoils, sliceIndex = 10)
 end
 
 ## B0 Maps (Assumes we have a B0 map from gradient echo scan named b0)
@@ -165,7 +165,7 @@ params[:correctionMap] = ComplexF32.(-1im .* resizedB0[:, :, selectedSlice])
 
 #totalRecon = sum(abs2,reco.data,dims=5)
 @info "Plotting Reconstruction \n"
-plotReconstruction(cartesianReco, 1:length(selectedSlice), resizedB0[:, :, selectedSlice], isSliceInterleaved = true, rotateAngle = 270)
+plot_reconstruction(cartesianReco, 1:length(selectedSlice), resizedB0[:, :, selectedSlice], isSliceInterleaved = true, rotateAngle = 270)
 
 ## Plot the image edges (feature comparison)
 
