+++ legacy/SpiralRecon_Cleaned_Mar2022_Human_SingleIntlv.jl
@@ -13,14 +13,14 @@ include("../utils/Utils.jl")
 ## ----------------------------- User-defined Variables -------------------------- ##
 
 ## Set true if we need to reload raw data compulsively.
-reloadSpiralData = true
+reload_spiral_data = true
 reloadGIRFData = true
 
 # Choose Slice (can be [single number] OR [1,2,3,...])
 # Leave empty ([]) to later select all slices
-sliceChoice = []; # TODO: read from ISMRMRD itself
+slice_choice = []; # TODO: read from ISMRMRD itself
 #[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15] # For multi-slice
-# sliceChoice = [6] # For single-slice
+# slice_choice = [6] # For single-slice
 
 
 ## Gyromagnetic ratio, in unit of Hz
@@ -29,9 +29,9 @@ gamma = 42577478
 
 
 ## Choose diffusion direction; starting from 0 (b=0) to the total number in MDDW protocol, e.g. for 6 diffusion directions, 1-6 stands for 6 DWIs)
-diffusionDirection = selector[:dif]
+diffusion_direction = selector[:dif]
 idxAverage = selector[:avg]
-nDiffusionDirections = params_general[:nDiffusionDirections] # TODO: Read from ISMRMRD itself
+num_total_diffusion_directions = params_general[:num_total_diffusion_directions] # TODO: Read from ISMRMRD itself
 
 # Which interleave to be reconstructed. For single-interleave data, it will always be set as 1; for multi-interleave data, the value set here will be used.
 # For multi-interleaved data, this value is ranging from [1:TotNumIntlv] (total number of interleaves), indicating which interleave to be reconstructed
@@ -39,27 +39,27 @@ startIndexIntlv = selector[:seg]
 
 
 ## Determine to reconstruct single-interleave data, or one interleave out of multi-interleave data.
-isDataSingleIntlv = isa(params_general[:fullPathScan], String)
+isDataSingleIntlv = isa(params_general[:scan_fullpath], String)
 
 
 ## ------------------------------------------------ Calculation Starts Here ---------------------------------------------------------- ##
 
-if params_general[:do_load_maps] && isfile(params_general[:fullPathSaveB0]) # # TODO ask for sense map (but split in magn/phase)
-    @info "Loading SENSE and B0 maps from $(params_general[:fullPathSaveSense]) and $(params_general[:fullPathSaveB0])"
+if params_general[:do_load_maps] && isfile(params_general[:b0_map_save_fullpath]) # # TODO ask for sense map (but split in magn/phase)
+    @info "Loading SENSE and B0 maps from $(params_general[:sensitivity_save_fullpath]) and $(params_general[:b0_map_save_fullpath])"
     # load maps, permute slice, sice files have geometric slice order
-    b0Maps = load_map(params_general[:fullPathSaveB0])
+    b0_maps = load_map(params_general[:b0_map_save_fullpath])
 
-    nSlices = size(b0Maps, 3)
-    sliceIndexArray = get_slice_order(nSlices, isSliceInterleaved = true)
+    num_slices = size(b0_maps, 3)
+    slice_index_array = get_slice_order(num_slices, isSliceInterleaved = true)
 
-    b0Maps = b0Maps[:, :, invperm(sliceIndexArray)]
-    cartesian_sensitivity = load_map(params_general[:fullPathSaveSense]; doSplitPhase = true)[:, :, invperm(sliceIndexArray), :]
+    b0_maps = b0_maps[:, :, invperm(slice_index_array)]
[…]
