+++ legacy/ReconConfig_SPIDI_0009.jl
@@ -18,30 +18,30 @@ params_general[:do_save_processed_map_scan] = false; # save ISMRMD file of prepr
 # params_general[:recon_id] = "2022-10-20_09_07_07"
 params_general[:recon_id] = "v2";
 params_general[:do_correct_with_b0_map] = true
-params_general[:doCorrectWithGIRFkxyz] = true
-params_general[:doCorrectWithGIRFk0] = true
+params_general[:do_correct_with_girf_k1] = true
+params_general[:do_correct_with_girf_k0] = true
 
-params_general[:nVirtualCoils] = 8;
-params_general[:doCoilCompression] = false;
-params_general[:fovShift] = [0, 0];# [0, -20]; # Unit: number of voxels
+params_general[:num_virtual_coils] = 8;
+params_general[:do_coil_compression] = false;
+params_general[:fov_shift] = [0, 0];# [0, -20]; # Unit: number of voxels
 
 ## Scan parameters, Additional acquisition information, e.g., slice distance etc.
-params_general[:sliceDistanceFactor_percent] = 000 # 400
+params_general[:slice_distance_factor_percent] = 000 # 400
 
 #Total number of ADC points BEFORE the rewinder at the end of the spiral readout. For gradient 508, use 15655 (out of 16084); for gradient 511, use 15445 (out of 15624).
-params_general[:numADCSamples] = 15445 # 15655
+params_general[:num_adc_samples] = 15445 # 15655
 # Matrix size of the reconstructed image. For gradient 508 with all 4 interleaves, use 200 for high resolution image; otherwise consider using 112 or 84 for a lower resolution. The FOV is 220 mm for both gradients 508 and 511.
-params_general[:reconSize] = (112, 112) #(112, 112) #(200, 200)
-params_general[:nReconIterations] = 20; # number of recon iterations (for both Cartesian and Spiral recon)
-params_general[:b0mapSmoothBeta] = 0.1 # for estimateB0Maps, * `β` - Regularization parameter controlling roughness penalty (larger = smoother, default 5e-4)
-params_general[:doNormalizeRecon] = false # set max abs to 1
-params_general[:scalingFactorSaveRecon] = 1.0e9 # 1 # typical range of recon intensities is 1e-7, rescale when saving, e.g., to 0...1000 roughly for fMRI analysis
+params_general[:recon_size] = (112, 112) #(112, 112) #(200, 200)
+params_general[:num_recon_iterations] = 20; # number of recon iterations (for both Cartesian and Spiral recon)
+params_general[:b0_map_beta] = 0.1 # for estimate_b0_maps, * `β` - Regularization parameter controlling roughness penalty (larger = smoother, default 5e-4)
+params_general[:do_normalize_recon] = false # set max abs to 1
+params_general[:saving_scalefactor] = 1.0e9 # 1 # typical range of recon intensities is 1e-7, rescale when saving, e.g., to 0...1000 roughly for fMRI analysis
 
 # Data selector
 #  Choose diffusion direction; starting from 0 (b=0) to the total number in MDDW protocol, e.g. for 6 diffusion directions, 1-6 stands for 6 DWIs)
-# boolean isCalledFromReconLoopGlobal is true, if this RunReconLoop is active
-# If isCalledFromReconLoopGlobal is false or not defined, the data selector needs to be defined here.
-if !(@isdefined isCalledFromReconLoopGlobal) || !isCalledFromReconLoopGlobal
+# boolean is_called_from_global_recon is true, if this RunReconLoop is active
[…]
