+++ docs/make.jl
@@ -9,7 +9,7 @@ lit = joinpath(@__DIR__, "lit", "examples")
 gen = joinpath(@__DIR__, "src", "generated")
 
 ipath_script = joinpath(lit, "joss_demo.jl")
-ipath_config = joinpath(lit, "ReconConfig_joss_demo.jl")
+ipath_config = joinpath(lit, "recon_config_joss_demo.jl")
 opath = gen
 Literate.markdown(ipath_script, opath; documenter = true)
 Literate.markdown(ipath_config, opath; documenter = true)
@@ -20,7 +20,7 @@ makedocs(
     pages = [
         "Home" => "index.md",
         "Utilities" => "Utilities.md",
-        "Examples" => [joinpath("generated/", "joss_demo.md"), joinpath("generated/", "ReconConfig_joss_demo.md")],
+        "Examples" => [joinpath("generated/", "joss_demo.md"), joinpath("generated/", "recon_config_joss_demo.md")],
     ],
 )
 
