### 2023-11-09 8ce0b68
Create CONTRIBUTING.md
