+++ CONTRIBUTING.md
@@ -1,0 +1,48 @@
+# Contributing to GIRFReco.jl
+We love your input! We want to make contributing to this project as easy and transparent as possible, whether it's:
+
+- Reporting a bug
+- Discussing the current state of the code
+- Submitting a fix
+- Proposing new features
+
+## We Develop with Github
+We use github to host code, to track issues and feature requests, as well as accept pull requests.
+
+We actively welcome your pull requests:
+
+1. Fork or clone the repo and create your branch from `main`.
+2. If you've added code that should be tested, add tests.
+3. Add docstrings to facilitate our generated documentation.
+4. Ensure your (and our) tests pass
+5. Post the PR or file an issue and we will work with you
+
+## Any contributions you make will be under the MIT Software License
+In short, when you submit code changes, your submissions are understood to be under the same [MIT License](http://choosealicense.com/licenses/mit/) that covers the project. Feel free to contact the maintainers if that's a concern.
+
+## Report bugs using Github's [issues](https://github.com/BRAIN-TO/GIRFReco.jl/issues)
+We use GitHub issues to track public bugs. Report a bug by [opening a new issue](https://github.com/BRAIN-TO/GIRFReco.jl/issues/new/choose); it's that easy!
+
+## Write bug reports with detail, background, and sample code
+
+**Great Bug Reports** tend to have:
+
+- A quick summary and/or background
+- Steps to reproduce
+  - Be specific!
+  - Give sample code if you can, a minimum working example (MWE) is best but we understand MRI involves complicated and large data
+- What you expected would happen
+- What actually happens
+- Notes (possibly including why you think this might be happening, or stuff you tried that didn't work)
+
+People *love* thorough bug reports. I'm not even kidding.
+
+## Use a Consistent Coding Style
+* We (mostly) use Julia [Blue](https://github.com/invenia/BlueStyle)
+* It would be great if you did as well
+
+## License
+By contributing, you agree that your contributions will be licensed under its MIT License.
+
+## References
+This document was adapted from the open-source contribution guidelines for [Facebook's Draft](https://github.com/facebook/draft-js/blob/a9316a723f9e918afde44dea68b5f9f39b7d9b00/CONTRIBUTING.md)
