### 2023-11-09 #5877 @alexjaffray
@Kevin-Mattheus-Moerman I've added it in the latest commit as CONTRIBUTING.md. Let me know if this is alright. It's https://github.com/BRAIN-TO/GIRFReco.jl/pull/4/commits/8ce0b68e94723417585ced4db1a6455d53769e4e
