### 2023-11-21 #5877 @Kevin-Mattheus-Moerman
@alexjaffray yes the contributing guidelines look fine now. Any other updates? Let us know when the reviewers can pick this up again.
