### 2023-12-12 0e89fe9
updated CONTRIBUTING to match our BSD 3-Clause license
