+++ CONTRIBUTING.md
@@ -17,8 +17,9 @@ We actively welcome your pull requests:
 4. Ensure your (and our) tests pass
 5. Post the PR or file an issue and we will work with you
 
-## Any contributions you make will be under the MIT Software License
-In short, when you submit code changes, your submissions are understood to be under the same [MIT License](http://choosealicense.com/licenses/mit/) that covers the project. Feel free to contact the maintainers if that's a concern.
+## Any contributions you make will be under the BSD 3-Clause Software License
+In short, when you submit code changes, your submissions are understood to be under the same [BSD 3-Clause License
+](https://choosealicense.com/licenses/bsd-3-clause/) that covers the project. Feel free to contact the maintainers if that's a concern.
 
 ## Report bugs using Github's [issues](https://github.com/BRAIN-TO/GIRFReco.jl/issues)
 We use GitHub issues to track public bugs. Report a bug by [opening a new issue](https://github.com/BRAIN-TO/GIRFReco.jl/issues/new/choose); it's that easy!
@@ -42,7 +43,7 @@ People *love* thorough bug reports. I'm not even kidding.
 * It would be great if you did as well
 
 ## License
-By contributing, you agree that your contributions will be licensed under its MIT License.
+By contributing, you agree that your contributions will be licensed under its BSD 3-Clause License.
 
 ## References
 This document was adapted from the open-source contribution guidelines for [Facebook's Draft](https://github.com/facebook/draft-js/blob/a9316a723f9e918afde44dea68b5f9f39b7d9b00/CONTRIBUTING.md)
