### 2023-12-19 BRAIN-TO/GIRFReco.jl#7 @alexjaffray
We have reviewed the issue and it will take some time to regenerate the figure, however we will keep in mind to follow a pixel-preserving image generation pipeline in the future. In the broader list of issues identified by the review, it is minor and does not really reduce the scientific value of the submitted work, so we will address it once the other issues are resolved.
