### 2023-12-21 BRAIN-TO/GIRFReco.jl#12 @felixhorger
Option 1 sounds good to me!
Using `RawAcquisitionData` and `AcquisitionData` is fine because MRIReco.jl made this kind of a standard.
One point that I nonetheless want to flag is the `utils` folder (as the name suggests it's functionality GIRFReco.jl needs but doesn't fit into a clear category):
- [ ] utils/Utils.jl: this file contains quite diverse functionality, you already noticed that because you named it "Utils", i.e. it's hard to find a matching name (contains e.g. plotting, manipulating trajectory and raw data, high-level functionality for applying GIRF corrections). I think this should be easy to split into separate files within GIRFReco.jl. For example utils/plot.jl, ismrmrd.jl, and utils/applyGIRF.jl.
- [ ] utils/fieldMapEstimator.jl: I suggest you outsource that because it is quite self-contained (apart from `estimateB0Maps()` which has to remain in GIRFReco.jl) and it would be nice to have this functionality outside of GIRFReco.jl. This functionality probably has to go somewhere else eventually (MRIReco?), but for now I propose it's alright to put it in a separate package. Is there any barrier to this I'm not seeing?
- [ ] utils/variationalSmoother.jl: Same as before, it's self-contained and could go into another package, until the opportunity arises to incorporate it into e.g. MRIReco's MRICoilSensitivities.jl.

I think doing that won't be a great effort (~1h?), and doesn't disagree with the major target of providing the pipeline as the main functionality but would have a positive effect on new users/developers of GIRFReco.jl.

Happy holidays everyone! :)
