### 2023-12-26 BRAIN-TO/GIRFReco.jl#12 @cncastillo
Hi! I hope everyone had a good Christmas 🎄.

> We have elected to pursue Option 1 as suggested in Issue https://github.com/BRAIN-TO/GIRFReco.jl/issues/14 for reducing the dependencies and also to streamline the running of the examples. This will achieve the target of a significantly more lightweight GIRFReco.jl.

Option 1 sounds good to me too.

> There are already many subpackages in the MRI Julia ecosystem; our repo has tried the best to utilize these subpackages, and it may **not be necessary for further compartmentalization**.

Regarding the package separation, I agree with @felixhorger that B0- and coil-related functions should not be in this package, and as said, the purpose of the package remains intact.

Cheers!
