### 2024-01-12 4f0be7d
allow option for raw return from merge_raw_interleaves
