+++ src/GIRFReco.jl
@@ -22,7 +22,28 @@ using FileIO
 using MRIFiles
 using MosaicViews
 
-include("../utils/utils.jl")
-include("../io/gradient_reader.jl")
+export plot_reconstruction,
+    plot_sense_maps,
+    calculate_b0_maps,
+    get_slice_order,
+    sync_traj_and_data!,
+    do_k0_correction!,
+    adjust_header!,
+    check_acquisition_nodes!,
+    validate_siemens_mrd!,
+    validate_acq_data!,
+    preprocess_cartesian_data,
+    remove_oversampling!,
+    merge_raw_interleaves,
+    apply_girf!,
+    apply_k0!,
+    save_map,
+    load_map,
+    shift_kspace!,
+    estimate_b0_maps,
+    read_gradient_text_file
+
+include("utils/utils.jl")
+include("io/gradient_reader.jl")
 
 end # module
