+++ src/utils/fieldmap_estimator.jl
@@ -1,4 +1,3 @@
-export estimate_b0_maps
 
 using Flux
 
