### 2024-01-12 d919eb3
update docstring and remove old TODO
