### 2024-01-12 5a90ed2
add 3d grad read-in support
