+++ src/GIRFReco.jl
@@ -1,26 +1,25 @@
 module GIRFReco
 
-using HDF5
-using MRIReco
-using LinearAlgebra
-using Dierckx
-using DSP
-using FourierTools
-using ImageBinarization
-using ImageEdgeDetection
-using Printf
-using ROMEO
-using NIfTI
-using Unitful
-using AxisArrays
-using ImageUtils
-using Plots
-using MRIGradients
-using DelimitedFiles
-using MAT
-using FileIO
-using MRIFiles
-using MosaicViews
+# using HDF5
+# using LinearAlgebra
+# using Dierckx
+# using DSP
+# using FourierTools
+# using ImageBinarization
+# using ImageEdgeDetection
+# using Printf
+# using ROMEO
+# using NIfTI
+# using Unitful
+# using AxisArrays
+# using ImageUtils
+# using Plots
+# using MRIGradients
+# using DelimitedFiles
+# using MAT
+# using FileIO
+# using MRIFiles
+# using MosaicViews
 
 export plot_reconstruction,
     plot_sense_maps,
