+++ docs/Project.toml
@@ -13,4 +13,5 @@ MRIGradients = "6ea382ba-d04a-44a2-b811-544bd2aee867"
 MRIReco = "bdf86e05-2d2b-5731-a332-f3fe1f9e047f"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 RegularizedLeastSquares = "1e9c538a-f78c-5de5-8ffb-0b6dbe892d23"
