### 2024-01-12 91803b6
remove NIfTI
