### 2024-01-12 3718126
add Plots to GIRFReco and remove fom demo
