+++ src/GIRFReco.jl
@@ -14,6 +14,7 @@ using Unitful
 using AxisArrays
 using ImageUtils
 using Plots
+using PlotlyJS
 using MRIGradients
 using FileIO
 using MRIBase
