+++ src/GIRFReco.jl
@@ -14,7 +14,7 @@ using Unitful
 using AxisArrays
 using ImageUtils
 using Plots
-using PlotlyJS
+# using PlotlyJS
 using MRIGradients
 using FileIO
 using MRIBase
