+++ recon/cartesian_recon.jl
@@ -136,7 +136,6 @@ end
 if params_general[:do_plot_recon]
     @info "Plotting Cartesian Results (Sensitivity Maps and B0 Maps)"
     # plot_sense_maps(sensitivity,num_coils)
-    plotlyjs()
     plot_reconstruction(cartesian_reco[:, :, :, 1], 1:size(cartesian_reco, 3), b0_maps, is_slice_interleaved = false, rotation = 180)
 end
 
