+++ Project.toml
@@ -8,46 +8,27 @@ AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
 DSP = "717857b8-e6f2-59f4-9121-6e50c889abd2"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Dierckx = "39dd38d3-220a-591b-8e3c-4c3a8c710a94"
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
-Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
 FourierTools = "b18b359b-aebc-45ac-a139-9c0ccbb2871e"
-HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
-HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
-ImageBinarization = "cbc4b850-ae4b-5111-9e64-df94c024a13d"
-ImageEdgeDetection = "2b14c160-480b-11ea-1b58-656063328ff7"
-ImageMorphology = "787d08f9-d448-5407-9aad-5290dd7ab264"
 ImageUtils = "8ad4436d-4835-5a14-8bce-3ae014d2950b"
-Images = "916415d5-f1e6-5110-898d-aaa5f9f070e0"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-LinearOperators = "5c8ed15e-5a4c-59e4-a42b-c7e8811fb125"
-Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 MAT = "23992714-dd62-5051-b70f-ba57cb901cac"
 MRIBase = "f7771a9a-6e57-4e71-863b-6e4b6a2f17df"
-MRICoilSensitivities = "c57eb701-aafc-44a2-a53c-128049758959"
-MRIFieldmaps = "5af58b90-8a55-4daf-adb4-2008a66a38cf"
 MRIFiles = "5a6f062f-bf45-497d-b654-ad17aae2a530"
 MRIGradients = "6ea382ba-d04a-44a2-b811-544bd2aee867"
-MRIOperators = "fb1137e3-90a6-46ce-a672-6e1e53d120f2"
-MRIReco = "bdf86e05-2d2b-5731-a332-f3fe1f9e047f"
 Measures = "442fdcdd-2543-5da2-b0f3-8c86c306513e"
 MosaicViews = "e94cdb99-869f-56ef-bcf0-1ae2bcbe0389"
 NFFT = "efe261a4-0d2b-5849-be55-fc731d526b0d"
 NFFTTools = "7424e34d-94f7-41d6-98a0-85abaf1b6c91"
-NIfTI = "a3a9e032-41b5-5fc4-967a-a6b7a19844d3"
-PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-PolygonInbounds = "e4521ec6-8c1d-418e-9da2-b3bc4ae105d6"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ROMEO = "1ea8258b-a15d-4adb-ad20-01632f467a84"
-RegularizedLeastSquares = "1e9c538a-f78c-5de5-8ffb-0b6dbe892d23"
 Scratch = "6c6a2e73-6563-6170-7368-637461726353"
[…]
