### 2024-01-12 BRAIN-TO/GIRFReco.jl#12 @alexjaffray
We have worked on this in #14 and have been able to remove quite a few dependencies, and are now in the process of consolidating our changes with the documentation workflow and writing tests for CI and coverage. We have also elected to remove the variational smoothing code from the repo. More or less finished, closing for now until we confirm that we can further reduce dependencies.
