### 2024-01-15 5ddfcf5
Added new version of the function get_slice_order
