+++ Project.toml
@@ -43,10 +43,7 @@ Dierckx = "0.5.3"
 FFTW = "1.7.1"
 FileIO = "1.16.1"
 FourierTools = "0.4.2"
-HDF5 = "0.16.15"
-ImageMorphology = "0.3.2"
 ImageUtils = "0.2.9"
-Images = "0.25.3"
 MAT = "0.10.5"
 MRIBase = "0.3.3"
 MRIFiles = "0.1.6"
@@ -55,7 +52,6 @@ Measures = "0.3.2"
 MosaicViews = "0.3.4"
 NFFT = "0.13.3"
 NFFTTools = "0.2.6"
-NIfTI = "0.5.9"
 Plots = "1.38.16"
 ROMEO = "1.0.1"
 Scratch = "1.2.0"

[…]
