### 2024-01-16 1a363e1
Fixed parameters for get_slice_order.
