### 2024-01-16 #5877 @alexjaffray
@Kevin-Mattheus-Moerman we have been working to incorporate the latest comments from the reviewers here: https://github.com/BRAIN-TO/GIRFReco.jl/pull/4
