### 2024-01-19 b434859
Setup environment for UT and CI
