+++ test/runtests.jl
@@ -1,5 +1,9 @@
 using GIRFReco
 using Test
+using Scratch
+
+global const tmpdir = @get_scratch!("temp")
+@info "For the output of tests, please refer to the directory $tmpdir"
 
 include("testutils.jl")
 include("testintegration.jl")
