+++ test/Project.toml
@@ -1,0 +1,7 @@
+[deps]
+GIRFReco = "cf1bd817-9e2d-403d-b1e7-15ef5a867bae"
+MRIReco = "bdf86e05-2d2b-5731-a332-f3fe1f9e047f"
+MRISampling = "9be66c26-f988-4649-80fc-f4721a4a33f2"
+MRISimulation = "8988da37-ea20-4fa6-9af7-8a6f6f9a8970"
+Scratch = "6c6a2e73-6563-6170-7368-637461726353"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
