### 2024-01-20 1088317
Added test function for save_map and load_map
