+++ test/runtests.jl
@@ -1,5 +1,6 @@
 using GIRFReco
 using MRISimulation
+using MRIReco
 using ImageUtils
 using Test
 using Scratch
