### 2024-01-25 8a9e3b1
add test for check_acquisition_nodes!
