### 2024-01-25 b64aeae
fix acquisition node test
