### 2024-01-26 ea7cf1c
add tests for JOSS paper
