+++ test/runtests.jl
@@ -4,6 +4,7 @@ using MRIReco
 using ImageUtils
 using Test
 using Scratch
+using Statistics
 
 global const tmpdir = @get_scratch!("temp")
 @info "For the output of tests, please refer to the directory $tmpdir"
