### 2024-01-26 3bd9226
add test for b0 map naive implementation
