### 2024-01-27 #5877 @alexjaffray
@Kevin-Mattheus-Moerman we are working through the test checklist here, and have implemented most of the other changes requested by the reviewers here https://github.com/BRAIN-TO/GIRFReco.jl/pull/4#issuecomment-1832685488. I imagine we will be ready to complete the above in 2 weeks.
