### 2024-01-29 a3289fd
add test for get_slice_order
