### 2024-01-29 d11f544
add test for sync_traj_and_data!
