### 2024-01-29 a491656
update docstring for apply_x! methods
