+++ docs/Project.toml
@@ -7,6 +7,7 @@ GIRFReco = "cf1bd817-9e2d-403d-b1e7-15ef5a867bae"
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
 ImageUtils = "8ad4436d-4835-5a14-8bce-3ae014d2950b"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 MRIBase = "f7771a9a-6e57-4e71-863b-6e4b6a2f17df"
 MRICoilSensitivities = "c57eb701-aafc-44a2-a53c-128049758959"
