### 2024-01-30 9fcb213
update integration test for girf applier
