+++ test/Project.toml
@@ -1,6 +1,7 @@
 [deps]
 GIRFReco = "cf1bd817-9e2d-403d-b1e7-15ef5a867bae"
 ImageUtils = "8ad4436d-4835-5a14-8bce-3ae014d2950b"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 MRIReco = "bdf86e05-2d2b-5731-a332-f3fe1f9e047f"
 MRISampling = "9be66c26-f988-4649-80fc-f4721a4a33f2"
 MRISimulation = "8988da37-ea20-4fa6-9af7-8a6f6f9a8970"
