### 2024-01-30 d94b218
add test for application of k0 and k1 girf
