### 2024-01-30 4eeb778
remove old gradient reader test and add test to suite
