+++ test/runtests.jl
@@ -11,5 +11,6 @@ using JLD2
 global const tmpdir = @get_scratch!("temp")
 @info "For the output of tests, please refer to the directory $tmpdir"
 
+include("testio.jl")
 include("testutils.jl")
 include("testintegration.jl")
