### 2024-01-30 7075253
Updated documentation workflo due to recent changes in JOSS review. Nowow the website source is the public repo.
