+++ docs/src/Utilities.md
@@ -1,4 +1,4 @@
 ```@autodocs
 Modules = [GIRFReco]
-Pages   = ["utils.jl", "gradient_reader.jl"]
+Pages   = ["utils.jl", "gradient_reader.jl", "fieldmap_estimator.jl"]
 ```
