+++ .github/workflows/documentation.yml
@@ -34,7 +34,7 @@ jobs:
         with:
           version: '1.9'
       - name: Install dependencies
-        run: julia --project=docs -e 'using Pkg; Pkg.add("MRIGradients"); Pkg.develop(PackageSpec(path=pwd())); Pkg.add("Documenter"); Pkg.add("Literate"); Pkg.instantiate()'
+        run: julia --project=docs -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
       - name: Build Documents through Documenter.jl
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # For authentication with GitHub Actions token
