### 2024-01-31 8613669
CompatHelper: bump compat for MRIBase to 0.4, (keep existing compat)
