### 2024-01-31 349767a
CompatHelper: bump compat for MRIFiles to 0.3, (keep existing compat)
