### 2024-01-31 17fa425
Fix descriptions of docstrings and function def in utils.
