+++ src/utils/utils.jl
@@ -1,25 +1,17 @@
-
-## Choose plotting backend to be PlotlyJS!
-# plotlyjs()
-
-## General Plotting function for the reconstruction
-
-# "Mosaic-plots reconstruction for selected slices and corresponding B0 map"
 """
-    plot_reconstruction(images, slices_index, b0; fig_handles = [], is_slice_interleaved = false)
+    plot_reconstruction(images, slices_index, b0; fig_handles = [], is_slice_interleaved = false, rotation = 0)
 Plots the magnitude and phase of the reconstructed images for a given slice or slices, along with a B₀ map if applicable
 
 # Arguments
 * `images` - Complex-valued images reconstructed using MRIReco.jl
 * `slices_index::Vector{Int}` - slices to plot
 * `b0` - off-resonance map to plot along with images
-* `fig_handles` - String vectors in size of [3,1] for titles of three figures (Magnitude & Phase of reconstructed images, and B0 maps)
+* `fig_handles` - String vectors in size of [3,1] for titles of three figures (Magnitude & Phase of reconstructed images, and B₀ maps)
 * `is_slice_interleaved::Bool` - for 2D scanning, indicate this value as `true` to make sure the slice order on the displayed results is correct
 * `rotation::Int` - Counterclock-wise rotation angle for each slice, should be a value from 0, 90, 180, 270 degrees
 """
-function plot_reconstruction(images, slices_index, b0; fig_handles = [], is_slice_interleaved = false, rotation = 0)
-    # plot()
-    ## If we need to re-order all slices
+function plot_reconstruction(images, slices_index, B₀; fig_handles = [], is_slice_interleaved = false, rotation = 0)
+
     num_slices = length(slices_index)
     reordered_slice_indices = zeros(Int16, size(slices_index))
 
@@ -35,15 +27,6 @@ function plot_reconstruction(images, slices_index, b0; fig_handles = [], is_slic
         error("rotation must be 0, 90, 180 or 270 degrees.")
     end
 
-    # Plot magnitude images (normalize)
-    # if length(fig_handles) < 1
-    #     figure("Magnitude Images")
-    # else
-    #     figure(fig_handles[1])
-    # end
-
-    # clf()
-
     magnitude_images = mapslices(x -> abs.(x) ./ maximum(abs.(x)), images[:, :, reordered_slice_indices], dims = [1, 2])
     if rotation == 90
         magnitude_images = mapslices(x -> rotr90(x), magnitude_images, dims = [1, 2])
@@ -55,7 +38,6 @@ function plot_reconstruction(images, slices_index, b0; fig_handles = [], is_slic
     magnitude_mosaic = mosaicview(magnitude_images, nrow = Int(floor(sqrt(num_slices))), npad = 5, rowmajor = true, fillvalue = 0)
 
     Plots.heatmap(magnitude_mosaic, show = true, plot_title = "|Images|", plot_titlevspan = 0.1, color = :grays, aspectratio = :equal)
-    # display(plot!())
 
     phase_images = angle.(images[:, :, reordered_slice_indices, 1, 1])
     if rotation == 90
@@ -68,20 +50,7 @@ function plot_reconstruction(images, slices_index, b0; fig_handles = [], is_slic
     phase_mosaic = mosaicview(phase_images, nrow = Int(floor(sqrt(num_slices))), npad = 5, rowmajor = true, fillvalue = 0)
 
     Plots.heatmap(phase_mosaic, show = true, plot_title = "∠ Images", plot_titlevspan = 0.1, color = :plasma, aspectratio = :equal)
-    #colorbar()
-    # display(plot!())
 
[…]
