### 2024-01-31 53d6e41
Fix descriptions in joss_demo.jl.
