+++ docs/lit/examples/joss_demo.jl
@@ -16,12 +16,6 @@ The configuration file is [`recon_config_joss_demo.jl`](@__REPO_ROOT_URL__/docs/
 The necessary Julia packages needed for spiral reconstruction.
 =#
 
-#Base packages for computation
-# using HDF5, LinearAlgebra, Dierckx, DSP, FourierTools, RegularizedLeastSquares, ImageUtils, PolygonInbounds
-
-#Packages for figure displaying
-# using MosaicViews, Plots, Images
-
 #Our developed packages
 using GIRFReco, MRIGradients
 
@@ -40,13 +34,12 @@ using PlotlyJS, Plots
 The following file, [`recon_config_joss_demo.jl`](@__REPO_ROOT_URL__/docs/lit/examples/recon_config_joss_demo.jl),
 includes general configuration for spiral reconstruction.
 It is necessary to execute this file to make sure all parameters are loaded.
-Sample Data that works with this script can be found at: https://doi.org/10.5281/zenodo.7779044
+Sample Data that works with this script can be found [here](https://doi.org/10.5281/zenodo.7779044).
 Please download, extract and set the root_project_path as the top level folder (should be something like /your/path/here/data-2, I've renamed mine to SPIDI)
 
 =#
 
-# root_project_path = "Your/Extracted/Data/Folder" # Root path of the data extracted from Zenodo
-root_project_path = "/srv/data/ajaffray/TORONTO_COLLAB/data/joss_data_zenodo"
+root_project_path = "[Your/Extracted/Data/Folder]" # Root path of the data extracted from Zenodo
 include("recon_config_joss_demo.jl")
 
 plotlyjs()
@@ -57,7 +50,7 @@ reload_girf_data = true # Set true if we need to reload GIRF data compulsively.
 
 #=
 Choose Slice ([single number] OR [1,2,31,...]）
-Leave empty ([]) or remove this line to later select all slices
+Leave an empty array ([]) or remove this line to select all slices
 =#
 slice_choice = []
 
@@ -75,19 +68,19 @@ is_single_interleave = isa(params_general[:scan_fullpath], String)
 
 #=
 Choose which interleave to be reconstructed. 
-For multi-interleave data, the range of this value is [1:num_total_interleaves] (total number of interleaves)
+For multi-interleave data, the range of this value is [1:num_total_interleaves].
 For single-interleave data, it should always be set as 1; for multi-interleave data, the value set here will be used, indicating which interleaves to be merged and reconstructed.
 =#
-start_idx_interleave = 1
+start_idx_interleave = 1;
 
 #===================================================
 ## 3. Image Reconstruction
 
 The steps of image reconstruction starts here.
 
-### 3.1 Calculation of B0 and Coil Sensitivity Maps
+### 3.1 Calculation of B₀ and Coil Sensitivity Maps
 
-The first step in reconstruction pipeline is to calculate the off-resonance (B0) maps `b0_map` 
+The first step in reconstruction pipeline is to calculate the off-resonance (B₀) maps `b0_map` 
 and coil sensitivity maps `cartesian_sensitivity` through the Cartesian reconstruction script 
 [cartesian_recon.jl](@__REPO_ROOT_URL__/recon/cartesian_recon.jl). 
 Ideally this script is execute once and the calculated maps are 
@@ -96,7 +89,7 @@ This is controlled by `do_load_maps` in general parameters.
 ===================================================#
 
 if params_general[:do_load_maps] && isfile(params_general[:b0_map_save_fullpath])
[…]
