### 2024-01-31 afcef98
fix b0 variable name and call to plot_sense_maps
