+++ docs/lit/examples/joss_demo.jl
@@ -35,8 +35,7 @@ The following file, [`recon_config_joss_demo.jl`](@__REPO_ROOT_URL__/docs/lit/ex
 includes general configuration for spiral reconstruction.
 It is necessary to execute this file to make sure all parameters are loaded.
 Sample Data that works with this script can be found [here](https://doi.org/10.5281/zenodo.7779044).
-Please download, extract and set the root_project_path as the top level folder (should be something like /your/path/here/data-2, I've renamed mine to SPIDI)
-
+Please download, extract and set the `root_project_path` as the top level folder (should be something like `/your/path/here/data`)
 =#
 
 root_project_path = "[Your/Extracted/Data/Folder]" # Root path of the data extracted from Zenodo
@@ -44,15 +43,15 @@ include("recon_config_joss_demo.jl")
 
 plotlyjs()
 
-# Two user defined parameters, just for this script.
-reload_spiral_data = true # Set true if we need to reload raw data compulsively.
-reload_girf_data = true # Set true if we need to reload GIRF data compulsively.
+# Two parameters determining whether we want to reload the raw data of spiral and GIRF.
+reload_spiral_data = true; # Set true if we need to reload raw data compulsively.
+reload_girf_data = true; # Set true if we need to reload GIRF data compulsively.
 
 #=
 Choose Slice ([single number] OR [1,2,31,...]）
 Leave an empty array ([]) or remove this line to select all slices
 =#
-slice_choice = []
+slice_choice = [];
 
 #=
 Choose which diffusion directions and averages to be processed. 
@@ -68,7 +67,9 @@ is_single_interleave = isa(params_general[:scan_fullpath], String)
 
 #=
 Choose which interleave to be reconstructed. 
-For multi-interleave data, the range of this value is [1:num_total_interleaves].
+
+For multi-interleave data, the range of this value is [1:num\_total\_interleaves].
+
 For single-interleave data, it should always be set as 1; for multi-interleave data, the value set here will be used, indicating which interleaves to be merged and reconstructed.
 =#
 start_idx_interleave = 1;
@@ -82,7 +83,8 @@ The steps of image reconstruction starts here.
 
 The first step in reconstruction pipeline is to calculate the off-resonance (B₀) maps `b0_map` 
 and coil sensitivity maps `cartesian_sensitivity` through the Cartesian reconstruction script 
-[cartesian_recon.jl](@__REPO_ROOT_URL__/recon/cartesian_recon.jl). 
+[cartesian_recon.jl](@__REPO_ROOT_URL__/recon/cartesian_recon.jl).
+
 Ideally this script is execute once and the calculated maps are 
 saved into files, which are loaded for future usage to save calculation time. 
 This is controlled by `do_load_maps` in general parameters. 
@@ -92,7 +94,6 @@ if params_general[:do_load_maps] && isfile(params_general[:b0_map_save_fullpath]
     @info "Loading SENSE and B₀ maps from $(params_general[:sensitivity_save_fullpath]) and $(params_general[:b0_map_save_fullpath])"
     cartesian_sensitivity = load_map(params_general[:sensitivity_save_fullpath]; do_split_phase = true)
     b0_maps = load_map(params_general[:b0_map_save_fullpath])
-    
     num_slices = size(b0_maps, 3)
 else
     @info "Running cartesian_recon to retrieve maps (cartesian_sensitivity and b0_maps)"
@@ -140,13 +141,17 @@ Slice 1, Slice 2 ... Slice N   Slice 1, Slice 2 ... Slice N     Slice 1, Slice 2
[…]
