+++ docs/lit/examples/recon_config_joss_demo.jl
@@ -7,7 +7,7 @@
 
 This recon_config.jl file describes all general reconstruction parameters, 
 as well as data locations and selections for an iterative non-Cartesian reconstruction that relies 
-on an external reference scan (Cartesian) to estimate calibration maps (coil sensitivities, B0 maps).
+on an external reference scan (Cartesian) to estimate calibration maps (coil sensitivities, B₀ maps).
 
 All parameters are stored in a dictionary named `params_general`.
 
@@ -18,17 +18,17 @@ params_general = Dict{Symbol,Any}();
 params_general[:gamma] = 42577478;
 
 # General options for reconstruction script:
-params_general[:do_load_maps] = true              # if true, reloads B0/SENSE maps instead of recalculating
-params_general[:do_save_recon] = true              # if true, saves reconstruction and all auxiliary image data (maps) as NIfTI files
-params_general[:do_plot_recon] = true              # if true, plots intermediate debugging and output recon figures (needs graphics, not recommended in multi-thread mode due to PyPlot)
-params_general[:do_process_map_scan] = true         # if true, compute sensitivity and B0 maps from reconstructed Cartesian scan   
+params_general[:do_load_maps] = true                 # if true, reloads B₀/SENSE maps instead of recalculating
+params_general[:do_save_recon] = true                # if true, saves reconstruction and all auxiliary image data (maps) as NIfTI files
+params_general[:do_plot_recon] = true                # if true, plots intermediate debugging and output recon figures (needs graphics, not recommended in multi-thread mode due to PyPlot)
+params_general[:do_process_map_scan] = true          # if true, compute sensitivity and B₀ maps from reconstructed Cartesian scan   
 params_general[:do_save_processed_map_scan] = false  # save ISMRMD file of preprocessed Cartesian data (before recon)
-params_general[:recon_id] = "v7"               # unique identifier for the saved result files
+params_general[:recon_id] = "v7"                     # unique identifier for the saved result files
 params_general[:do_correct_with_b0_map] = true       # whether perform off-resonance correction
-params_general[:do_correct_with_girf_k1] = true    # whether perform 1st order GIRF correction
-params_general[:do_correct_with_girf_k0] = true    # whether perform 1st order GIRF correction
-params_general[:do_coil_compression] = false       # whether perform coil compression
-params_general[:do_normalize_recon] = false       # if true, set the range of magnitude image as [0 1]
+params_general[:do_correct_with_girf_k1] = true      # whether perform 1st order GIRF correction
+params_general[:do_correct_with_girf_k0] = true      # whether perform 1st order GIRF correction
+params_general[:do_coil_compression] = false         # whether perform coil compression
+params_general[:do_normalize_recon] = false;         # if true, set the range of magnitude image as [0 1]
 
 # General parameters for reconstruction:
 params_general[:num_adc_samples] = 15650             # total number of ADC points BEFORE the rewinder at the end of the spiral readout. Need to check with data. # 15504 for 523, 15655 for gradient 508
@@ -39,7 +39,8 @@ params_general[:saving_scalefactor] = 1.0e9    # typical range of recon intensit
 params_general[:num_virtual_coils] = 8                 # if perform coil compression, the number of coils to be compressed to
[…]

[…]
