### 2024-01-31 669034b
Changed the template path of the downloaded files from Zenodo.
