### 2024-02-07 #5877 @alexjaffray
@Kevin-Mattheus-Moerman @cncastillo @mathieuboudreau @felixhorger we have finished with our revisions as of commit https://github.com/BRAIN-TO/GIRFReco.jl/commit/d3d59f7d2d189a38c5540278aa9e961e44781538. We have implemented most of the changes suggested and we feel it has resulted in a much improved package. We are very grateful for the feedback we received and have incorporated it. We have also implemented changes in MRIGradients to improve it alongside this submission.

We have decided to tackle the off-resonance estimation and the figure resolution issue as a next step outside of the scope of the review, but are looking forward to addressing them soon as well.

We look forward to hearing from all of you.
Best, Alex, Tim, and Lars.
