### 2024-02-16 #5877 @cncastillo
Sure, congratulations to the authors for all the work!

For my part I will re-read the paper, and try to reproduce the results. If I remember correctly the in-vivo data is not publicly available, so I will try with the phantom data.
