### 2024-02-17 BRAIN-TO/GIRFReco.jl#25 @felixhorger
plot_sense_maps missing arg
Missed the `num_channels` argument
https://github.com/BRAIN-TO/GIRFReco.jl/blob/d3d59f7d2d189a38c5540278aa9e961e44781538/src/utils/utils.jl#L85-L93

Or maybe you meant to remove it but forgot in the demo: https://github.com/BRAIN-TO/GIRFReco.jl/blob/d3d59f7d2d189a38c5540278aa9e961e44781538/docs/lit/examples/joss_demo.jl#L248
