### 2024-02-18 BRAIN-TO/GIRFReco.jl#25 @felixhorger
Same for fig_handles in `plot_reconstruction()`
https://github.com/BRAIN-TO/GIRFReco.jl/blob/d3d59f7d2d189a38c5540278aa9e961e44781538/docs/lit/examples/joss_demo.jl#L343-L350
