### 2024-02-18 #5877 @felixhorger
I think the package is good to go once https://github.com/BRAIN-TO/GIRFReco.jl/issues/25 is fixed. I'll take your word for the interpolation issue with the figure 😄
Congrats and great job @alexjaffray @nbwuzhe @mrikasper, really appreciate the amount of extra work you put in, this is a valuable contribution!
