### 2024-02-19 09fe1dd
Fixing plot function argument mentioned in issue #25.
