### 2024-02-19 BRAIN-TO/GIRFReco.jl#25 @nbwuzhe
Hi @felixhorger thank you so much for pointing out these issues. I have fixed them in #26. We haven't figured out a way to setup test cases for plotting functions which leads to about 30% short of code coverage for now. Maybe we need to do that in the future.
