### 2024-02-19 BRAIN-TO/GIRFReco.jl#25 @alexjaffray
Closing as issue has been fixed
