### 2024-03-13 BRAIN-TO/GIRFReco.jl#27 @cncastillo
Portability problems when running phantom example
Hi everyone,

I hope everything is going well. I have tried to run the example and had a few problems. I will describe the problems I encountered in "chronological order", trying to describe the experience of a new user trying to run [this example](https://brain-to.github.io/GIRFReco.jl/dev/generated/joss_demo/).

To run the example in phantom data, I did not find the link to download the data within the joss_demo. For convenience, I would add a link there, so I don't need to return to the README. Preferably, I would not need to care about this; it should just download the data into `data/` in the current directory.

The second problem I encountered was where to put `recon_config_joss_demo.jl` and `data/` as it assumes a specific folder structure not described in the example (or at least not in the setup section). This could also be solved by downloading `recon_config_joss_demo.jl` into the current directory.

Finally I concluded that it needs to be like this:
```
data/
joss_demo.jl
recon_config_joss_demo.jl
```
Then, when I tried running it in a `temp` environment (clean env. like someone new to Julia), I realized that you are using mulitple packages:
```julia
using GIRFReco, MRIGradients
using MRIReco, FileIO, MRIFiles, MRIBase, MRICoilSensitivities
using RegularizedLeastSquares, Flux
using ImageTransformations
using PlotlyJS, Plots
```
Does someone that just want to run the example actually need to install all of them? When trying to install them, I had some disk space issues, but after deleting some files I managed to proceed. Instructions telling the user what packages to actually install (`add PackageX`) in the environment must be included, as GIRFReco has a lot of those inside now. This could be a simple `Project.toml`, as adding all of them manually is a little cumbersome.

Now, where I had some real problems is when it tried to run `recon/cartesian_recon.jl` that is also searching for `"../src/utils/fieldmap_estimator.jl"`, similar to #9. This is a big portability problem of the example, as it assumes that I am running it within the repo. I am unsure where I need to run the example, maybe this assumes the folder structure before the JOSS changes (?). I could have downloaded the repo and run it there, but I believe this is an incorrect approach, and the user shouldn't need to do this. If the user needs additional functions to run an example, those need to be included in the package.

This is unfortunately as far as I got. I think the example needs to be refined, so the user can download `joss_demo.jl` and its `Project.toml` into a new folder, activate the `Project.toml` with `] activate .` to install the necessary packages, run `include("joss_demo.jl")` in the REPL, and it should just work. These steps need to be described in [the example](https://brain-to.github.io/GIRFReco.jl/dev/generated/joss_demo/) with a warning telling the user that "X GB" will be downloaded for the phantom data. First impressions are very important for new users (especially new Julia users), and the provided example needs to be flawless.

Hope this is helpful,
Cheers!
