### 2024-03-14 BRAIN-TO/GIRFReco.jl#27 @mrikasper
Dear Carlos @cncastillo,

Oh wow, sorry for all the issues when trying to run the example! I agree that this should have been a flawless experience for you (and any other user), and I am a bit confused why we didn't catch this earlier when going through all the feedback from the first round of reviews.

I thought we did quite some refactoring to have, for example, a specific [Julia environment for the docs](https://github.com/BRAIN-TO/GIRFReco.jl/blob/main/docs/Project.toml) (and example), but maybe we are missing something fundamental here. Or could it be something as as simple as not having updated the `README` consistently?

For example, would a "Getting started" section like this work for the questions about the environments/packages to install?

```
using Pkg
Pkg.add("GIRFReco")

# activating example environment
cd docs
Pkg.activate(".")
Pkg.instantiate()

# run example
cd lit/examples
julia joss_demo.jl
```

One thing I am not sure about is how to navigate to the `docs` folder if users can be anywhere on their path.

Maybe the reason we did not spot this has to do with us having multiple (development) versions of the code and a specific IDE configuration on our machines instead of a clean slate? Just to double-check: You ran the current version of the code from the `main` branch?

Regarding the user experience, how would you go about making the example data more readily available? The original data is very large (as is raw MR data in general, unfortunately), but we could think about making only a subset of this available in a smaller ISMRMRD file. For downloading it automatically, how would you recommend implementing this in Julia? Via testing artifacts?

Again, so sorry for your difficulties, we will try to fix this soon, and any feedback to my questions here would be highly appreciated.

Thank you so much for your help!

All the best,
Lars
