### 2024-03-15 BRAIN-TO/GIRFReco.jl#27 @cncastillo
> Or could it be something as as simple as not having updated the README consistently?

I think it could be as simple as that, to update the instructions and add links to what needs to be downloaded. But that should be in https://brain-to.github.io/GIRFReco.jl/dev/generated/joss_demo/, not in the README. The example needs to be self-contained! There is no purpose to [joss_demo](https://brain-to.github.io/GIRFReco.jl/dev/generated/joss_demo/), if I need to keep going back to the README to make the example work.

> You ran the current version of the code from the main branch?

I am using the registered GIRFReco v0.1.1. I did not clone the repo or go to the package's location after the installation, as that is something a new user wouldn't do (the [instructions](https://github.com/BRAIN-TO/GIRFReco.jl#quick-installation) say only to type `] add GIRFReco`).

> Regarding the user experience, how would you go about making the example data more readily available? The original data is very large (as is raw MR data in general, unfortunately), but we could think about making only a subset of this available in a smaller ISMRMRD file. For downloading it automatically, how would you recommend implementing this in Julia? Via testing artifacts?

I don't think the size of the data is really a problem, so I think it is okay to keep it as it is (with the corresponding warning in the demo). I am not completely sure about the artifacts. Just make sure that it is not downloaded when doing `add GIRFReco`. To download files programmatically, you can use `Downloads.download`.  For the Zenodo data you could require doing it differently, but if the user is instructed in joss_demo to download the data (with a link) and unzip it following a certain folder structure, it should be fine.

I was imagining something like this at the beginning of `joss_demo.jl`:
```julia
# Getting necessary files
using Downloads
Downloads.download(example_zenodo_data_url, "./data.zip") #Then uncompress
Downloads.download(example_config_url, "./recon_config_joss_demo.jl")
Downloads.download(example_env_url, "./Project.toml")
# Unzip data
using ZipFile
unzip("./data.zip", exdir="./data")
# Setting environment
using Pkg
Pkg.activate(".") #Maybe something else is necessary here
```
I believe downloading data, a config file, Project.toml, and the actual example are reasonable, but not `recon/` or the whole repo.

> Now, where I had some real problems is when it tried to run recon/cartesian_recon.jl that is also searching for "../src/utils/fieldmap_estimator.jl", similar to https://github.com/BRAIN-TO/GIRFReco.jl/issues/9.

Another option, is to include files that are in the package with (**do not** `cd` to package):
```julia
include(joinpath(dirname(pathof(GIRFReco)), "../recon/cartesian_recon.jl"))
```
but it is putting the problem under the rug, if those functions are necessary to display the functionalities of the package, export them in the package or bundle them with the example. If you believe this is not the most basic example, you can simplify it. The user shouldn't need to run code inside the package.
