### 2024-03-18 BRAIN-TO/GIRFReco.jl#27 @alexjaffray
@cncastillo thanks for pointing these issues out and we can work towards solving them. The demo should be, as you said, self-contained.

I like your proposed solution using the Downloads and Zipfile approach to managing our external data location, we will proceed that way.

Regarding `"./recon/cartesian_recon.jl"`, I will functionalize it properly, it makes sense to export it as a function instead of a script as suggested in #9.
