### 2024-03-19 9d81967
Update README.md; added Getting Started section inspired by MRIReco
