### 2024-03-22 2cac86b
update example setup
