+++ example/recon_config_joss_demo.jl
@@ -1,0 +1,153 @@
+#-----------------------------------------------------------------------------------
+# # [GIRFReco.jl Example Configuration](@id example_config)
+#-----------------------------------------------------------------------------------
+
+#=
+## Introduction
+
+This recon_config.jl file describes all general reconstruction parameters, 
+as well as data locations and selections for an iterative non-Cartesian reconstruction that relies 
+on an external reference scan (Cartesian) to estimate calibration maps (coil sensitivities, B₀ maps).
+
+All parameters are stored in a dictionary named `params_general`.
+
+=#
+params_general = Dict{Symbol,Any}();
+
+# Gyromagnetic ratio, in unit of Hz
+params_general[:gamma] = 42577478;
+
+# General options for reconstruction script:
+params_general[:do_load_maps] = true                 # if true, reloads B₀/SENSE maps instead of recalculating
+params_general[:do_save_recon] = true                # if true, saves reconstruction and all auxiliary image data (maps) as NIfTI files
+params_general[:do_plot_recon] = true                # if true, plots intermediate debugging and output recon figures (needs graphics, not recommended in multi-thread mode due to PyPlot)
+params_general[:do_process_map_scan] = true          # if true, compute sensitivity and B₀ maps from reconstructed Cartesian scan   
+params_general[:do_save_processed_map_scan] = false  # save ISMRMD file of preprocessed Cartesian data (before recon)
+params_general[:recon_id] = "v7"                     # unique identifier for the saved result files
+params_general[:do_correct_with_b0_map] = true       # whether perform off-resonance correction
+params_general[:do_correct_with_girf_k1] = true      # whether perform 1st order GIRF correction
+params_general[:do_correct_with_girf_k0] = true      # whether perform 1st order GIRF correction
+params_general[:do_coil_compression] = false         # whether perform coil compression
+params_general[:do_normalize_recon] = false;         # if true, set the range of magnitude image as [0 1]
+
+# General parameters for reconstruction:
+params_general[:num_adc_samples] = 15650             # total number of ADC points BEFORE the rewinder at the end of the spiral readout. Need to check with data. # 15504 for 523, 15655 for gradient 508
+params_general[:recon_size] = [200, 200, 1]          # the matrix size of the reconstructed images. Needs to specify 1 on Z dimension for 2D images
+params_general[:num_recon_iterations] = 40             # number of recon iterations (for both Cartesian and Spiral recon)
+params_general[:b0_map_beta] = 0.1             # for estimate_b0_maps, * `β` - Regularization parameter controlling roughness penalty (larger = smoother, default 5e-4)
+params_general[:saving_scalefactor] = 1.0e9    # typical range of recon intensities is 1e-7, rescale when saving, e.g., to 0...1000 roughly for fMRI analysis
+params_general[:num_virtual_coils] = 8                 # if perform coil compression, the number of coils to be compressed to
+params_general[:fov_shift] = [-2, -5]               # amount of FOV shift; in unit of number of voxels in [x,y] direction (for our in-vivo dataset this is [0, -15], change as necessary)
+params_general[:slice_distance_factor_percent] = 400 # Scan parameters, Additional acquisition information, e.g., slice distance etc.
+params_general[:num_total_diffusion_directions] = 6;          # Need to specify total diffusion directions included in the raw data
+
+#=
+
[…]
