+++ example/joss_demo.jl
@@ -1,0 +1,354 @@
+#-----------------------------------------------------------------------------------
+# # [GIRFReco.jl Example Script](@id example_script)
+#-----------------------------------------------------------------------------------
+
+#=
+This page demonstrates an example script for using GIRFReco.jl
+
+This page was generated from the following Julia file: [`joss_demo.jl`](@__REPO_ROOT_URL__/docs/lit/examples/joss_demo.jl)
+
+The configuration file is [`recon_config_joss_demo.jl`](@__REPO_ROOT_URL__/docs/lit/examples/recon_config_joss_demo.jl)
+=#
+
+#=
+## 1. Setup
+
+The necessary Julia packages needed for spiral reconstruction.
+=#
+
+# mktempdir(parent=tempdir(); prefix="jl_", cleanup=true) -> path
+
+#Our developed packages
+using GIRFReco, MRIGradients
+
+#MRIReco and its sub-packages
+using MRIReco, FileIO, MRIFiles, MRIBase, MRICoilSensitivities
+
+using RegularizedLeastSquares, Flux
+
+using ImageTransformations
+
+using PlotlyJS, Plots
+
+#=
+## 2. Configurations for reconstruction
+
+The following file, [`recon_config_joss_demo.jl`](@__REPO_ROOT_URL__/docs/lit/examples/recon_config_joss_demo.jl),
+includes general configuration for spiral reconstruction.
+It is necessary to execute this file to make sure all parameters are loaded.
+Sample Data that works with this script can be found [here](https://doi.org/10.5281/zenodo.7779044).
+Please download, extract and set the `root_project_path` as the top level folder (should be something like `/your/path/joss_data_zenodo/`)
+=#
+
+root_project_path = "/your/path/joss_data_zenodo/" # Root path of the data extracted from Zenodo
+include("recon_config_joss_demo.jl")
+
+plotlyjs()
+
+# Two parameters determining whether we want to reload the raw data of spiral and GIRF.
+reload_spiral_data = true; # Set true if we need to reload raw data compulsively.
+reload_girf_data = true; # Set true if we need to reload GIRF data compulsively.
+
+#=
+Choose Slice ([single number] OR [1, 2, 31, ...]）
+Leave an empty array ([]) or remove this line to select all slices
+=#
+slice_choice = [];
+
+#=
+Choose which diffusion directions and averages to be processed. 
+Diffusion direction index starts from 0 (b=0) to the total number in MDDW protocol (e.g. for 6 diffusion directions, 1-6 stands for 6 DWIs). 
+Index for average starts from 1.
+=#
+diffusion_direction = 0
+idx_average = 1
+num_total_diffusion_directions = params_general[:num_total_diffusion_directions]
+
+## Determine to reconstruct single-interleave data, or one interleave out of multi-interleave data.
+is_single_interleave = isa(params_general[:scan_fullpath], String)
+
+#=
+Choose which interleave to be reconstructed. 
+
+For multi-interleave data, the range of this value is [1:num\_total\_interleaves].
+
+For single-interleave data, it should always be set as 1; for multi-interleave data, the value set here will be used, indicating which interleaves to be merged and reconstructed.
+=#
+start_idx_interleave = 1;
+
+#===================================================
+## 3. Image Reconstruction
+
+The steps of image reconstruction starts here.
+
+### 3.1 Calculation of B₀ and Coil Sensitivity Maps
+
+The first step in reconstruction pipeline is to calculate the off-resonance (B₀) maps `b0_map` 
[…]
