+++ src/GIRFReco.jl
@@ -40,9 +40,11 @@ export plot_reconstruction,
     load_map,
     shift_kspace!,
     estimate_b0_maps,
-    read_gradient_text_file
+    read_gradient_text_file,
+    run_example
 
 include("utils/utils.jl")
 include("io/gradient_reader.jl")
+include("utils/example_setup.jl") # convenience file for example
 
 end # module
