+++ example/Project.toml
@@ -1,0 +1,21 @@
+[deps]
+FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
+GIRFReco = "cf1bd817-9e2d-403d-b1e7-15ef5a867bae"
+HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
+ImageTransformations = "02fcd773-0e25-5acc-982a-7f6622650795"
+ImageUtils = "8ad4436d-4835-5a14-8bce-3ae014d2950b"
+Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
+MRIBase = "f7771a9a-6e57-4e71-863b-6e4b6a2f17df"
+MRICoilSensitivities = "c57eb701-aafc-44a2-a53c-128049758959"
+MRIFiles = "5a6f062f-bf45-497d-b654-ad17aae2a530"
+MRIGradients = "6ea382ba-d04a-44a2-b811-544bd2aee867"
+MRIReco = "bdf86e05-2d2b-5731-a332-f3fe1f9e047f"
+MRISimulation = "8988da37-ea20-4fa6-9af7-8a6f6f9a8970"
+PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
+RegularizedLeastSquares = "1e9c538a-f78c-5de5-8ffb-0b6dbe892d23"
+
+[compat]
+MRIReco = "0.7.1"

[…]
