+++ src/utils/example_setup.jl
@@ -1,8 +1,13 @@
 # Getting necessary files
+using Pkg
+
+Pkg.activate(".") #Maybe something else is necessary here
+Pkg.add("ZipFile")
+
+using Downloads
+using ZipFile
 
 function run_example()
-
-    using Downloads
 
     example_zenodo_data_url = "https://zenodo.org/records/7779045/files/data_girfreco_03_28_2023.zip?download=1"
     example_config_url = ""
@@ -19,12 +24,9 @@ function run_example()
     Downloads.download(example_env_url, "./Project.toml")
 
     # Unzip data
-    using ZipFile
     unzip("./data.zip", exdir="./data")
 
     # Setting environment
-    using Pkg
-    Pkg.activate(".") #Maybe something else is necessary here
 
     include("joss_demo.jl")
     include("fieldmap_estimator.jl")
