+++ Project.toml
@@ -8,6 +8,7 @@ AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
 DSP = "717857b8-e6f2-59f4-9121-6e50c889abd2"
 DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Dierckx = "39dd38d3-220a-591b-8e3c-4c3a8c710a94"
+Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 FourierTools = "b18b359b-aebc-45ac-a139-9c0ccbb2871e"
@@ -29,13 +30,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
-
-[extras]
-CPUSummary = "2a0fbf3d-bb9c-48f3-b0a9-814d99fd7ab9"
-FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
-
 [compat]
-julia = "1.9"
 AxisArrays = "0.4.6"
 DSP = "0.7.8"
 DelimitedFiles = "1.9.1"
@@ -55,6 +50,11 @@ NFFTTools = "0.2.6"
 Plots = "1.38.16"
 ROMEO = "1.0.1"
 Scratch = "1.2.0"
+Statistics = "1.9.0"
 StatsBase = "0.33.21"
 Unitful = "1.14.0"
-Statistics = "1.9.0"
+julia = "1.9"
+
+[extras]
+CPUSummary = "2a0fbf3d-bb9c-48f3-b0a9-814d99fd7ab9"
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
