### 2024-03-23 ed1cb05
add HTTP for zip file download and Downloads for .jls
