+++ example/joss_demo.jl
@@ -40,7 +40,7 @@ Sample Data that works with this script can be found [here](https://doi.org/10.5
 Please download, extract and set the `root_project_path` as the top level folder (should be something like `/your/path/joss_data_zenodo/`)
 =#
 
-root_project_path = "/your/path/joss_data_zenodo/" # Root path of the data extracted from Zenodo
+root_project_path = "./joss_data_zenodo" # Root path of the data extracted from Zenodo
 include("recon_config_joss_demo.jl")
 
 plotlyjs()
