### 2024-03-25 1db055f
add option for mid-example unzip
