### 2024-03-25 c5b6b84
remove ZipFile and Downloads deps
