+++ docs/lit/examples/joss_demo.jl
@@ -97,7 +97,7 @@ if params_general[:do_load_maps] && isfile(params_general[:b0_map_save_fullpath]
     num_slices = size(b0_maps, 3)
 else
     @info "Running cartesian_recon to retrieve maps (cartesian_sensitivity and b0_maps)"
-    include("../../../recon/cartesian_recon.jl")
+    cartesian_sensitivity, b0_maps = run_cartesian_recon(params_general)
     num_slices = size(b0_maps, 3)
 end
 
