### 2024-03-26 c365f45
Added info for data downloading script.
