+++ example/download_data.jl
@@ -6,11 +6,13 @@ example_zenodo_data_url = "https://zenodo.org/records/7779045/files/data_girfrec
 
 # Zip data file download
 if ~isfile("./data.zip")
+    @info "Downloading demo data from Zenodo."
     HTTP.download(example_zenodo_data_url, "./data.zip")
 end
 
 # Unzipping data files
 if ~isdir("./joss_data_zenodo")
+    @info "Unzipping demo data."
     mkpath("./joss_data_zenodo")
     run(`unzip data.zip -d ./joss_data_zenodo/`)
 end
