### 2024-03-27 ad5ab73
Release 0.1.2
New release after separating example scripts.
