### 2024-03-27 0e0ae5c
Added compat for Flux and MRIFieldmaps
Added compat for Flux and MRIFieldmaps
