+++ Project.toml
@@ -37,12 +37,14 @@ DSP = "0.7.8"
 DelimitedFiles = "1.9.1"
 Dierckx = "0.5.3"
 FFTW = "1.7.1"
+Flux = "0.14.0"
 FileIO = "1.16.1"
 FourierTools = "0.4.2"
 ImageUtils = "0.2.9"
 MAT = "0.10.5"
 MRIBase = "0.3.3"
 MRIFiles = "0.1.6"
+MRIFieldmaps = "0.0.1"
 MRIGradients = "0.1.2"
 Measures = "0.3.2"
 MosaicViews = "0.3.4"
