+++ example/download_data.jl
@@ -1,6 +1,7 @@
 using Pkg
 
 Pkg.activate(".")
+using HTTP
 
 example_zenodo_data_url = "https://zenodo.org/records/7779045/files/data_girfreco_03_28_2023.zip?download=1"
 
@@ -13,6 +14,5 @@ end
 # Unzipping data files
 if ~isdir("./joss_data_zenodo")
     @info "Unzipping demo data."
-    mkpath("./joss_data_zenodo")
-    run(`unzip data.zip -d ./joss_data_zenodo/`)
+    run(`unzip data.zip -d ./`)
 end
