+++ Project.toml
@@ -44,7 +44,7 @@ ImageUtils = "0.2.9"
 MAT = "0.10.5"
 MRIBase = "0.3.3"
 MRIFiles = "0.1.6"
-MRIFieldmaps = "0.0.1"
+MRIFieldmaps = "0.0.3"
 MRIGradients = "0.1.2"
 Measures = "0.3.2"
 MosaicViews = "0.3.4"
