+++ Project.toml
@@ -16,6 +16,7 @@ ImageUtils = "8ad4436d-4835-5a14-8bce-3ae014d2950b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MAT = "23992714-dd62-5051-b70f-ba57cb901cac"
 MRIBase = "f7771a9a-6e57-4e71-863b-6e4b6a2f17df"
+MRICoilSensitivities = "c57eb701-aafc-44a2-a53c-128049758959"
 MRIFieldmaps = "5af58b90-8a55-4daf-adb4-2008a66a38cf"
 MRIFiles = "5a6f062f-bf45-497d-b654-ad17aae2a530"
 MRIGradients = "6ea382ba-d04a-44a2-b811-544bd2aee867"
@@ -26,6 +27,7 @@ NFFTTools = "7424e34d-94f7-41d6-98a0-85abaf1b6c91"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 ROMEO = "1ea8258b-a15d-4adb-ad20-01632f467a84"
+RegularizedLeastSquares = "1e9c538a-f78c-5de5-8ffb-0b6dbe892d23"
 Scratch = "6c6a2e73-6563-6170-7368-637461726353"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
@@ -37,14 +39,14 @@ DSP = "0.7.8"
 DelimitedFiles = "1.9.1"
 Dierckx = "0.5.3"
 FFTW = "1.7.1"
-Flux = "0.14.0"
 FileIO = "1.16.1"
+Flux = "0.14.0"
 FourierTools = "0.4.2"
 ImageUtils = "0.2.9"
 MAT = "0.10.5"
 MRIBase = "0.3.3"
-MRIFiles = "0.1.6"
 MRIFieldmaps = "0.0.3"
+MRIFiles = "0.1.6"
 MRIGradients = "0.1.2"
 Measures = "0.3.2"
 MosaicViews = "0.3.4"
