+++ Project.toml
@@ -20,6 +20,7 @@ MRICoilSensitivities = "c57eb701-aafc-44a2-a53c-128049758959"
 MRIFieldmaps = "5af58b90-8a55-4daf-adb4-2008a66a38cf"
 MRIFiles = "5a6f062f-bf45-497d-b654-ad17aae2a530"
 MRIGradients = "6ea382ba-d04a-44a2-b811-544bd2aee867"
+MRIReco = "bdf86e05-2d2b-5731-a332-f3fe1f9e047f"
 Measures = "442fdcdd-2543-5da2-b0f3-8c86c306513e"
 MosaicViews = "e94cdb99-869f-56ef-bcf0-1ae2bcbe0389"
 NFFT = "efe261a4-0d2b-5849-be55-fc731d526b0d"
