+++ Project.toml
@@ -1,7 +1,7 @@
 name = "GIRFReco"
 uuid = "cf1bd817-9e2d-403d-b1e7-15ef5a867bae"
 authors = ["Tim Wu <tim.wuzhe@gmail.com>"]
-version = "0.1.2"
+version = "0.1.3"
 
 [deps]
 AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
@@ -46,15 +46,18 @@ FourierTools = "0.4.2"
 ImageUtils = "0.2.9"
 MAT = "0.10.5"
 MRIBase = "0.3.3"
+MRICoilSensitivities = "0.1.3"
 MRIFieldmaps = "0.0.3"
 MRIFiles = "0.1.6"
 MRIGradients = "0.1.2"
+MRIReco = "0.7.1"
 Measures = "0.3.2"
 MosaicViews = "0.3.4"
 NFFT = "0.13.3"
 NFFTTools = "0.2.6"
 Plots = "1.38.16"
 ROMEO = "1.0.1"
+RegularizedLeastSquares = "0.13"
 Scratch = "1.2.0"
 Statistics = "1.9.0"
 StatsBase = "0.33.21"
