+++ Project.toml
@@ -45,10 +45,10 @@ Flux = "0.14.0"
 FourierTools = "0.4.2"
 ImageUtils = "0.2.9"
 MAT = "0.10.5"
-MRIBase = "0.3.3"
-MRICoilSensitivities = "0.1.3"
+MRIBase = "0.3.1"
+MRICoilSensitivities = "0.1.1"
 MRIFieldmaps = "0.0.3"
-MRIFiles = "0.1.6"
+MRIFiles = "0.1.5"
 MRIGradients = "0.1.2"
 MRIReco = "0.7.1"
 Measures = "0.3.2"
