+++ Project.toml
@@ -57,7 +57,7 @@ NFFT = "0.13.3"
 NFFTTools = "0.2.6"
 Plots = "1.38.16"
 ROMEO = "1.0.1"
-RegularizedLeastSquares = "0.13"
+RegularizedLeastSquares = "0.9.1"
 Scratch = "1.2.0"
 Statistics = "1.9.0"
 StatsBase = "0.33.21"
