+++ example/cartesian_recon.jl
@@ -1,145 +1,0 @@
-include("fieldmap_estimator.jl")
-
-## Dictionary of frequently changed parameters
-# include("recon_config.jl")
-
-## Load data files
-
-@info "Loading Data Files"
-
-b0_filename = params_general[:map_scan_fullpath];
-
-# filename for preprocessed data (remove oversampling, permute dimensions wrt MRIReco)
-processed_filename = params_general[:processed_map_scan_fullpath]
-
-# Set the data file name (Change this for your own system)
-cartesian_data_file = ISMRMRDFile(b0_filename)
-
-# read in the raw data from the ISMRMRD file into a RawAcquisitionData object
-raw = RawAcquisitionData(cartesian_data_file)
-
-if params_general[:do_process_map_scan] 
-    # Preprocess Data and save!
-    preprocess_cartesian_data(raw::RawAcquisitionData, true; filename = processed_filename)
-
-end
-
-# Load preprocessed data!
-processed_cartesian_data_file = ISMRMRDFile(processed_filename)
-
-raw_new = RawAcquisitionData(processed_cartesian_data_file)
-cartesian_acq_data = AcquisitionData(raw_new, estimateProfileCenter = true)
-
-# Define coils and slices
-num_coils = size(cartesian_acq_data.kdata[1], 2)
-num_slices = numSlices(cartesian_acq_data)
-
-# Get the order of slices from the RawAcqData header
-slice_idx_array_cartesian = get_slice_order(raw, num_slices, 1, 1)
-
-# Define TEs 
-# Echo times for field map raw data, in ms
-TE1 = raw_new.params["TE"][1]
-TE2 = raw_new.params["TE"][2]
-
-# shift FOV to middle :) 
-#TODO: in MRIReco v0.7, try: correctOffset(cartesian_acq_data, [0 -20 0])
-shift_kspace!(cartesian_acq_data, params_general[:fov_shift]) # amount of FOV shift; in unit of number of voxels in [x,y] direction
-
-#changeFOV!(cartesian_acq_data,[1.5,1.5])
-
-## Don't have to recalculate sense maps for both scans but possibly it could make a
-#  difference in Diffusion scans
-
-@info "Calculating Sense Maps"
-
-# from reading docstring, min thresholding
-cartesian_sensitivity = espirit(cartesian_acq_data, (6, 6), 30, eigThresh_1 = 0.01, eigThresh_2 = 0.9)
-
-# from Lars (7T spirals)
-#cartesian_sensitivity = espirit(cartesian_acq_data,(6,6),30,eigThresh_1=0.02, eigThresh_2=0.98)
-# from Alexander (Phantom?)
-#cartesian_sensitivity = espirit(cartesian_acq_data,(4,4),12,eigThresh_1=0.01, eigThresh_2=0.98)
-
-# normalize for consistency with saving/loading and better ranges of reconstruction values
-cartesian_sensitivity /= maximum(abs.(cartesian_sensitivity))
-
-res_x = fieldOfView(cartesian_acq_data)[1] ./ size(cartesian_sensitivity)[1]
-res_y = fieldOfView(cartesian_acq_data)[2] ./ size(cartesian_sensitivity)[2]
-res_z = fieldOfView(cartesian_acq_data)[3] .* (1 + params_general[:slice_distance_factor_percent] ./ 100.0) # for 2D only, since FOV[3] is slice thickness then, but gap has to be observed
-resolution_mm = (res_x, res_y, res_z)
-
-# save SENSE maps
-if params_general[:do_save_recon] # TODO: include elements to save as tuple, e.g., ["b0", "sense", "recon"], same for load
-    # TODO: use correct slice order everywhere, e.g., when saving/loading maps for spiral recon
[…]
