+++ example/download_data.jl
@@ -1,6 +1,8 @@
 using Pkg
 
 Pkg.activate(".")
+Pkg.add("HTTP")
+Pkg.instantiate()
 using HTTP
 
 example_zenodo_data_url = "https://zenodo.org/records/7779045/files/data_girfreco_03_28_2023.zip?download=1"
