+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"EditURL = \"../../lit/examples/joss_demo.jl\"","category":"page"},{"location":"generated/joss_demo/#example_script","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"","category":"section"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"This page demonstrates an example script for using GIRFReco.jl","category":"page"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"This page was generated from the following Julia file: joss_demo.jl","category":"page"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"The configuration file is recon_config_joss_demo.jl","category":"page"},{"location":"generated/joss_demo/#.-Setup","page":"GIRFReco.jl Example Script","title":"1. Setup","text":"","category":"section"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"The necessary Julia packages needed for spiral reconstruction.","category":"page"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"#Our developed packages\nusing GIRFReco, MRIGradients\n\n#MRIReco and its sub-packages\nusing MRIReco, FileIO, MRIFiles, MRIBase, MRICoilSensitivities\n\nusing RegularizedLeastSquares, Flux\n\nusing ImageTransformations\n\nusing PlotlyJS, Plots","category":"page"},{"location":"generated/joss_demo/#.-Configurations-for-reconstruction","page":"GIRFReco.jl Example Script","title":"2. Configurations for reconstruction","text":"","category":"section"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"The following file, recon_config_joss_demo.jl, includes general configuration for spiral reconstruction. It is necessary to execute this file to make sure all parameters are loaded. Sample Data that works with this script can be found here. Please download, extract and set the root_project_path as the top level folder (should be something like /your/path/joss_data_zenodo/)","category":"page"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"root_project_path = \"/your/path/joss_data_zenodo/\" # Root path of the data extracted from Zenodo\ninclude(\"recon_config_joss_demo.jl\")\n\nplotlyjs()","category":"page"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"Two parameters determining whether we want to reload the raw data of spiral and GIRF.","category":"page"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"reload_spiral_data = true; # Set true if we need to reload raw data compulsively.\nreload_girf_data = true; # Set true if we need to reload GIRF data compulsively.\nnothing #hide","category":"page"},{"location":"generated/joss_demo/","page":"GIRFReco.jl Example Script","title":"GIRFReco.jl Example Script","text":"Choose Slice ([single number] OR [1, 2, 31, ...]） Leave an empty array ([]) or remove this line to select all slices","category":"page
[…]
