### 2024-04-02 c55a3dc
Improving plot configs
