### 2024-04-02 #5877 @mathieuboudreau
Green light from me! All done with the review checklist - I had not issues downloading the package and running the example inside inside Visual Studio Code with the Julia extension.

The project is very well documented and the paper reads well also.

Well done!
