### 2024-04-02 #5877 @alexjaffray
@mathieuboudreau thank you for trying the package and for taking the time to review it!

@cncastillo @felixhorger the most recent release (v0.1.5 or main branch) of GIRFReco.jl should fix the issues @cncastillo ran into previously running the exaxmple.
