### 2024-04-03 BRAIN-TO/GIRFReco.jl#27 @cncastillo
Hi! Thanks for all the hard work. It is much easier to run the example now :smile:  (`include("run_example.jl")`). I was able to replicate the results in the paper with ease.

My results:
![image](https://github.com/BRAIN-TO/GIRFReco.jl/assets/5957134/aacb81d7-bf03-4303-b4ae-8ceb7736d0fd)
Results in the paper:
![image](https://github.com/BRAIN-TO/GIRFReco.jl/assets/5957134/a3e4e18b-c9e6-4645-a54d-9bacc971180c)

Some minor comments:

-  If I delete `data.zip`, `download_data.jl` tries to download the data again, even if the data is already unzipped in `joss_data_zenodo`
- I was not able to set `slice_choice = [];` to anything else without having unexpected results (erros, or the reconstruction showed the wrong slices)
- The plots generated in Julia are a little stretched (not 1:1 aspect ratio). It was better to check the results in the produced NifTI files. On my NifTI viewer, scrolling through the slices looked weird (probably because of the diffusion dimension). I think a recommended viewer could be suggested.
- For convenience, I would paste part2 of the instructions in [the example's README.md](https://github.com/BRAIN-TO/GIRFReco.jl/blob/main/example/README.md) onto [the beginning of the example](https://brain-to.github.io/GIRFReco.jl/dev/generated/joss_demo/), for the docs to be self-contained (someone could have missed the additional README).
-  Maybe add some precomputed results in the docs? (like the one below without the GIRF corrections)
![image](https://github.com/BRAIN-TO/GIRFReco.jl/assets/5957134/6498900c-1c38-4dc3-a640-5257e4b933f9)

Congratulations again! You have addressed all my major comments :)
