### 2024-04-03 #5877 @cncastillo
Congratulations to the authors. The last release (GIRFReco.jl v0.1.5) fixed all my issues and I have successfully replicated the results in the paper. I have no further comments besides some minor ones in https://github.com/BRAIN-TO/GIRFReco.jl/issues/27.

Very well done!!
