### 2024-04-03 #5877 @Kevin-Mattheus-Moerman
@alexjaffray congratulations, it looks like all reviewers have given their blessing and ticked all the boxes. Do you think you can address those minor things pointed out in https://github.com/BRAIN-TO/GIRFReco.jl/issues/27 before we move on?
