### 2024-04-03 802a34f
Fixed unnecessary re-download after deleting zip file after extracting all data.
