### 2024-04-03 #5877 @alexjaffray
@Kevin-Mattheus-Moerman we are happy to address the items pointed out in https://github.com/BRAIN-TO/GIRFReco.jl/issues/27, @nbwuzhe, @mrikasper and myself will do this and let you know once it's complete so that we can move forward. Thanks for your time in facilitating this review over the past months.

Thanks @cncastillo @felixhorger @mathieuboudreau for the insightful and helpful reviews and for the guidance through the review process.
