+++ docs/lit/examples/joss_demo.jl
@@ -3,12 +3,54 @@
 #-----------------------------------------------------------------------------------
 
 #=
-This page demonstrates an example script for using GIRFReco.jl
+This page gives a step-by-step demonstration of an example script for using GIRFReco.jl, which consists of the following two files:
 
-This page was generated from the following Julia file: [`joss_demo.jl`](@__REPO_ROOT_URL__/docs/lit/examples/joss_demo.jl)
+- A Julia script of an example pipeline for spiral reconstruction: [`joss_demo.jl`](@__REPO_ROOT_URL__/example/joss_demo.jl)
 
-The configuration file is [`recon_config_joss_demo.jl`](@__REPO_ROOT_URL__/docs/lit/examples/recon_config_joss_demo.jl)
+- A configuration file which defines parameters for the spiral reconstruction: [`recon_config_joss_demo.jl`](@__REPO_ROOT_URL__/example/recon_config_joss_demo.jl)
+
+## 0. Quick Start
+
+To have a quick run of the example, simply take the following steps.
+
+### (1) Cloning the package
+
+Download or clone the repo to your local by:
+
+```
+git clone git@github.com:BRAIN-TO/GIRFReco.jl.git
+```
+
+Then enter the `example` folder:
+
+```
+cd GIRFReco.jl/example
+```
+
+Alternatively, for the one who are using Visual Studio Code, simply open the `example` sub-folder from the Menu option `File -> Open Folder...`.
+
+
+### (2) Run in Julia REPL
+
+We recommend to use Visual Studio Code with Julia extension (steps of installation can be found [here](https://code.visualstudio.com/docs/languages/julia)) to avoid possible image displaying issue, especially for those using SSH and X11 forwarding.
+
+The Julia REPL can be launched from the Command Palette (For Windows and Linux: Press `Shift + Ctrl + P`; For Mac: Press `Shift + Command + P`) by searching the command `Julia: Start REPL`.
+
+In the launched REPL, simpley execute the following command after launching REPL to run the whole demonstration script (including data download):
+
+```julia
+include("run_example.jl")
+```
+
+If you only want to download the demonstration dataset, run:
+
+```julia
+include("download_data.jl")
+```
+
 =#
+
+
 
 #=
 ## 1. Setup
