### 2024-04-04 307a693
Added heuristics to README.
