### 2024-04-04 BRAIN-TO/GIRFReco.jl#27 @nbwuzhe
Hi @cncastillo , thanks (again) for your careful review of our code. This repo improved dramatically after taking the suggestions in this and your previous comments. Please accept my appreciation.

Regarding your comments:

Q1: If I delete data.zip, download_data.jl tries to download the data again, even if the data is already unzipped in joss_data_zenodo

A1: This has been fixed in the latest commits. Thanks for pointing out this issue.

Q2: I was not able to set slice_choice = []; to anything else without having unexpected results (erros, or the reconstruction showed the wrong slices)

A2: Really appreciate your careful test. We had a thorough test of multiple combinations of slice numbers & orders. This issue has been fixed in the latest commits.

Q3: The plots generated in Julia are a little stretched (not 1:1 aspect ratio). It was better to check the results in the produced NifTI files. On my NifTI viewer, scrolling through the slices looked weird (probably because of the diffusion dimension). I think a recommended viewer could be suggested.

A3: After surveying multiple backends of Plots, we chose PlotlyJS considering it's reaching a good balance between multiple factors. We have tried our best to achieve aspect ratio = 1 (we have specified this option in the `heatmap` call). As you are more experienced in Julia, do you have any further suggestions regarding image displaying (e.g. backend choice, configuration, etc.)

For now, we are only demonstrating our recon pipeline using b = 0 images, especially for multi-interleave recon (if [this line](https://github.com/BRAIN-TO/GIRFReco.jl/blob/f1beeb024a30eb919a7f94f1e6a07489d23ed1d5/example/recon_config_joss_demo.jl#L97) is activated), so there may be some displaying issues across the dimension of diffusion for some NIFTI viewers. Meanwhile, our dataset has a relatively large slice gap (8mm gap vs 2mm slice thickness) so the slice location may not look smooth. We were majorly using FSLeyes and I have suggested it in the README.

Q4: For convenience, I would paste part2 of the instructions in [the example's README.md](https://github.com/BRAIN-TO/GIRFReco.jl/blob/main/example/README.md) onto [the beginning of the example](https://brain-to.github.io/GIRFReco.jl/dev/generated/joss_demo/), for the docs to be self-contained (someone could have missed the additional README).

A4: This part is fulfilled. Thanks for the suggestion.

Q5: Maybe add some precomputed results in the docs? (like the one below without the GIRF corrections)

A5: We would like to give this opportunity to our readers/users to play around with the configurations of our example and see how the results change. I added a few words in the [README](https://github.com/BRAIN-TO/GIRFReco.jl/tree/main/example).
