+++ example/README.md
@@ -1,12 +1,12 @@
 # Quick Guide for Running GIRFReco Example
 
-You may find code details of the example script [here](https://brain-to.github.io/GIRFReco.jl).
+You can find code details of the example script [here](https://brain-to.github.io/GIRFReco.jl).
 
 ## System Configuration
 
-Though other lower versions may still be compatible, our testing environment is Julia with version 1.9.3, and we recommend to use the same or higher version of Julia.
+Though compatibility with older versions of Julia may exist, our testing environment is Julia with version 1.9.3, and we recommend to use the same or newer version of Julia.
 
-We also recommend to leave at least 30 GB disk space for the data of demonstration and all dependent Julia packages.
+We also recommend to leave at least 30 GB disk space for the example data and all dependent Julia packages.
 
 The reconstructed Cartesian and Spiral images in NIfTI format will be in the folder `joss_data_zenodo/results`. We recommend to use [FSLeyes](https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FSLeyes) as the viewer to open them.
 
@@ -35,7 +35,7 @@ We recommend to use Visual Studio Code with Julia extension (steps of installati
 
 The Julia REPL can be launched from the Command Palette (For Windows and Linux: Press `Shift + Ctrl + P`; For Mac: Press `Shift + Command + P`) by searching the command `Julia: Start REPL`.
 
-In the launched REPL, simpley execute the following command after launching REPL to run the whole demonstration script (including data download):
+In the launched REPL, simply execute the following command after launching the REPL to run the whole demonstration script (including the data download):
 
 ```julia
 include("run_example.jl")
@@ -47,14 +47,14 @@ If you only want to download the demonstration dataset, run:
 include("download_data.jl")
 ```
 
-### (3) After your first example running...
+### (3) After running your first example ...
 
 "Watch a movie, or be part of one". You are also encouraged to be part of our repo by playing around with the [configuration file]([`recon_config_joss_demo.jl`](./recon_config_joss_demo.jl)) and see how (and why) the results are changed. For example:
 
-(1) How the results change when we turn off the B₀ correction (by setting `do_correct_with_b0_map` as `false`)?
+(1) How do the results change when we turn off the B₀ correction (by setting `do_correct_with_b0_map` as `false`)?
 
 (2) What will the images look like when we turn off the zeroth and/or the first order GIRF correction (set `do_correct_with_girf_k0` and/or `do_correct_with_girf_k1` as `false`)?
 
-(3) Try reconstruction with multiple spiral interleaves by enabling line 97.
+(3) Try reconstruction with multiple spiral interleaves by enabling line 97. What changes do you see?
 
 And so many more options to explore...
