### 2024-04-04 cd68b6c
Increment Version and Update Author info for final release
