+++ Project.toml
@@ -1,7 +1,7 @@
 name = "GIRFReco"
 uuid = "cf1bd817-9e2d-403d-b1e7-15ef5a867bae"
-authors = ["Tim Wu <tim.wuzhe@gmail.com>"]
-version = "0.1.5"
+authors = ["Tim Wu <tim.wuzhe@gmail.com>", "Alexander Jaffray <alexander.jaffray@gmail.com>","Lars Kasper <lars.kasper@utoronto.ca>]
+version = "0.1.6"
 
 [deps]
 AxisArrays = "39de3d68-74b9-583c-8d2d-e117c070f3a9"
