### 2024-04-04 #5877 @alexjaffray
@Kevin-Mattheus-Moerman @cncastillo @mathieuboudreau @felixhorger thank you so much for all of the help during the review process, I think we have arrived at a much more complete body of work as a result of your feedback, review and efforts.

We have addressed the minor items in https://github.com/BRAIN-TO/GIRFReco.jl/issues/27 and have created a new release with the relevant fixes, so I think we are good to proceed. I think there remains a single plotting issue (aspect ratio squeezed) that we are asking @cncastillo directly about here (https://github.com/BRAIN-TO/GIRFReco.jl/issues/27#issuecomment-2036204588) and which should be resolved quickly. Otherwise, we are happy to move on.

Thanks @nbwuzhe and @mrikasper for their work on this package since the beginning and their mentorship throughout this project, it would not have been possible without our work as a team.
