### 2024-04-04 BRAIN-TO/GIRFReco.jl#27 @nbwuzhe
> > A3: After surveying multiple backends of Plots, we chose PlotlyJS considering it's reaching a good balance between multiple factors. We have tried our best to achieve aspect ratio = 1 (we have specified this option in the heatmap call). As you are more experienced in Julia, do you have any further suggestions regarding image displaying (e.g. backend choice, configuration, etc.)
>
> Looking at it again, the problem was not the aspect ratio of the original plot. The problem occurs when you zoom, as you can create any shape, distorting the image aspect ratio.
>
> I know how to fix this with `PlotlyJS.jl` [directly](https://plotly.com/julia/axes/#fixed-ratio-axes) (`yaxis=attr(scaleanchor="x", scaleratio=1)`). However, as `Plots.jl` wraps `PlotlyJS.jl`, I am unsure how to modify it. Based on this:
>
> https://stackoverflow.com/questions/69574843/zooming-in-does-not-adapt-date-axis-when-using-julias-plotlyjs-in-vscode
>
> It seems that could be a limitation of Plots.jl, you can try adding `ticks=:native`. To be honest, now that there is a recommended NifTI viewer, that should be enough.
>
> Congratulations again on the hard work! 😄

Hi @cncastillo really appreciate your insights.

I tried your suggestion using `ticks=:native` option. It has two issues with the current settings: (1) This option is in conflict with the `yflip` option we are using (they won't co-exist, complaining `KeyError: key :range not found`), which is quite essential to make sure the mosaic figure is displayed from top to bottom (the default will be from bottom to up, leaving upper right corner with blank subfigures when the columns of the last row are not filled). (2) Even if we removed `yflip` option, using `ticks=:native` will undo our efforts keeping the initial figure with aspect ratio = 1 (`aspect_ratio` option doesn't work anymore in this case). I will talk to @alexjaffray and @mrikasper to discuss whether it's a good idea to switch to native PlotlyJS to get rid of such a dilemma.

Kindly let us know if you have any further suggestions and we are happy to include them in our next release.

Best regards,

Tim
