### 2024-04-06 BRAIN-TO/GIRFReco.jl#27 @alexjaffray
@cncastillo @mrikasper @nbwuzhe the most recent push and release should fix everything but the plotting aspec ratio issue, and we now promote a NifTI viewer to see the results instead. Closing this.
