### 2024-04-26 #5877 @alexjaffray
@Kevin-Mattheus-Moerman I have completed the list above. The version is v0.1.7, which contains an updated affiliation, and I have archived it on Zenodo here (10.5281/zenodo.11075548). I have also made sure that the license is consistent across the archive and the software package (BSD-3 Clause "New" or "Revised" License)
