### 2024-04-30 #5877 @alexjaffray
@Kevin-Mattheus-Moerman the items with no DOI given are correct, there is no DOI available for those 5 items and we have cited them as the authors of the packages/platforms suggest.

Regarding the two items which have a "possibly valid" DOI, the dois suggested are correct but do not link directly to the works. Using the links we provided provides a direct link to those conference abstracts. What would you suggest in this case?
