### 2024-04-30 #5877 @mrikasper
@Alexander: For the two ISMRM Abstracts, I think it's more long-lasting to put the DOI, and the full document is only one click away then. Similar to how most publishers DOIs redirect to the HTML page with the paper abstract, not the full text or pdf.

Of course, if there is a second field for URL (besides DOI) you can use the full-text link there.

All the best,
Lars

--
Lars Kasper, PhD (he/him)
***@***.***

MR Physicist
Toronto Neuroimaging Facility (ToNI), Department of Psychology, University of Toronto

325 Huron St, Toronto, ON M5S 3J7
phone: +1 416-978-7613 (office) / +1 416-946-0356 (scanner)
