+++ paper/paper.bib
@@ -15,8 +15,8 @@
 	abstract = {We investigated the use of physically-motivated corrections for non-Cartesian imaging within a fast, open-source MR reconstruction framework written in the Julia programming language. Building on existing Julia libraries for MR image reconstruction, we developed a comprehensive, open-source pipeline for non-Cartesian image reconstruction including B0 inhomogeneity correction, gradient system characterization and eddy current correction. We demonstrated that spiral images can be reconstructed using this pipeline with high geometric accuracy in a fast, accessible and efficient manner on modest hardware.},
 	booktitle = {Proc. {Intl}. {Soc}. {Mag}. {Reson}. {Med}. 30},
 	author = {Jaffray, Alexander and Wu, Zhe and Uludağ, Kâmil and Kasper, Lars},
+	doi = {10.58530/2022/2435},
 	year = {2022},
-	keywords = {⛔ No DOI found},
 	pages = {2435},
 }
 
@@ -27,6 +27,7 @@
 	language = {en},
 	booktitle = {Proc. {Intl}. {Soc}. {Mag}. {Reson}. {Med}. 30},
 	author = {Wu, Zhe and Jaffray, Alexander and Vannesjo, S. Johanna and Uludağ, Kâmil and Kasper, Lars},
+	doi = {10.58530/2022/0641},
 	year = {2022},
 	pages = {0641},
 }
