### 2024-05-02 #5877 @Kevin-Mattheus-Moerman
@alexjaffray it looks like we are good to proceed. However, I would add one minor comment. On the Zenodo archive listing the spelling for the author Kâmil Uludağ does not feature the special characters. It may be good to edit this person's name there to match how it is listed in the paper too. You can do as if you wish but we can proceed to process this for acceptance at this point.
