### 2024-05-02 07eaa4c
Added CITATION.cff from JOSS Acceptance Recommendation
