### 2024-05-03 #5877 @Kevin-Mattheus-Moerman
@alexjaffray great. Well done and congratulations on this JOSS publication!

And a special thanks to the fantastic reviewers!!! @cncastillo, @mathieuboudreau, @felixhorger
